Plutarch.Builtin.Array

Documentation

newtype PArray (a :: S -> Type) (s :: S)

A packed collection of values.

Since: 1.11.0

Constructors

PArray (Term s (PArray a))

plengthOfArray :: forall (a :: S -> Type) (s :: S). Term s (PArray a :--> PInteger)

Get the length of an array, as per CIP-138.

Since: 1.11.0

plistToArray :: forall (a :: S -> Type) (s :: S). Term s (PBuiltinList a :--> PArray a)

Convert a (builtin) list to an array with the same contents in the same order, as per CIP-138.

Since: 1.11.0

pindexArray :: forall (a :: S -> Type) (s :: S). Term s (PArray a :--> (PInteger :--> a))

Index an array, as per CIP-138.

Since: 1.11.0