Plutarch.Builtin.Data

Documentation

newtype PData (s :: S)

Constructors

PData (Term s PData)

newtype PAsData (a :: S -> Type) (s :: S)

Constructors

PAsData (Term s a)

pchooseData :: forall (s :: S) (a :: S -> Type). Term s (PData :--> (a :--> (a :--> (a :--> (a :--> (a :--> a))))))

pasConstr :: forall (s :: S). Term s (PData :--> PBuiltinPair PInteger (PBuiltinList PData))

pasMap :: forall (s :: S). Term s (PData :--> PBuiltinList (PBuiltinPair PData PData))

plistData :: forall (s :: S). Term s (PBuiltinList PData :--> PData)

pasList :: forall (s :: S). Term s (PData :--> PBuiltinList PData)

pasInt :: forall (s :: S). Term s (PData :--> PInteger)

pasByteStr :: forall (s :: S). Term s (PData :--> PByteString)

pserialiseData :: forall (s :: S). Term s (PData :--> PByteString)

Serialise any builtin data to its cbor represented by a builtin bytestring

pconstrBuiltin :: forall (s :: S). Term s (PInteger :--> (PBuiltinList PData :--> PAsData (PBuiltinPair PInteger (PBuiltinList PData))))

data PBuiltinPair (a :: S -> Type) (b :: S -> Type) (s :: S)

A builtin Plutus pair.

Since: 1.12.0

Constructors

PBuiltinPair (Term s a) (Term s b)

pfstBuiltin :: forall (s :: S) (a :: S -> Type) (b :: S -> Type). Term s (PBuiltinPair a b :--> a)

Deprecated: Use pmatch instead

psndBuiltin :: forall (s :: S) (a :: S -> Type) (b :: S -> Type). Term s (PBuiltinPair a b :--> b)

Deprecated: Use pmatch instead

ppairDataBuiltin :: forall (s :: S) (a :: S -> Type) (b :: S -> Type). Term s (PAsData a :--> (PAsData b :--> PBuiltinPair (PAsData a) (PAsData b)))

Construct a builtin pair of PData elements.

Uses PAsData to preserve more information about the underlying PData.

data PBuiltinList (a :: S -> Type) (s :: S)

Plutus BuiltinList

Constructors

PCons (Term s a) (Term s (PBuiltinList a))

PNil

pheadBuiltin :: forall (s :: S) (a :: S -> Type). Term s (PBuiltinList a :--> a)

ptailBuiltin :: forall (s :: S) (a :: S -> Type). Term s (PBuiltinList a :--> PBuiltinList a)

pheadTailBuiltin :: forall (a :: S -> Type) (b :: S -> Type) (s :: S). Term s (PBuiltinList a) -> (Term s a -> Term s (PBuiltinList a) -> Term s b) -> Term s b

Use this in preference to pheadBuiltin and ptailBuiltin on the same PBuiltinList, as this will be faster. This is also faster than a pmatch, as the PNil case is omitted.

Since: 1.13.0

pchooseListBuiltin :: forall (s :: S) (a :: S -> Type) (b :: S -> Type). Term s (PBuiltinList a :--> (b :--> (b :--> b)))

pnullBuiltin :: forall (s :: S) (a :: S -> Type). Term s (PBuiltinList a :--> PBool)

pconsBuiltin :: forall (s :: S) (a :: S -> Type). Term s (a :--> (PBuiltinList a :--> PBuiltinList a))