Plutarch.Internal.Other

Documentation

printTerm :: forall (a :: S -> Type). HasCallStack => Config -> (forall (s :: S). Term s a) -> String

Prettyprint a Term via the PLC pretty printer

TODO: Heavily improve. It's unreadable right now.

We could convert the de Bruijn indices into names with:

show . prettyPlcReadableDef . (\(Right p) -> p) . Scripts.mkTermToEvaluate . compile $ term

printScript :: Script -> String

Prettyprint a compiled Script via the PLC pretty printer

data Flip (f :: k1 -> k2 -> Type) (a :: k2) (b :: k1)

Type level flip operation, reversing the order of arguments Commonly used in Plutarch to get the PTryFromExcess associated type of PTryFrom for a Plutarch type

Since: 1.12.0