Plutarch.Internal.PlutusType

Documentation

class PlutusType (a :: S -> Type)

type PInnermost (a :: S -> Type) = PInnermost' (PInner a) a

type family PlutusTypeStratConstraint strategy :: (S -> Type) -> Constraint

pcon' :: forall (s :: S). PlutusType a => a s -> Term s (PInner a)

pmatch' :: forall (s :: S) (b :: S -> Type). PlutusType a => Term s (PInner a) -> (a s -> Term s b) -> Term s b

pmatch :: forall a (s :: S) (b :: S -> Type). PlutusType a => Term s a -> (a s -> Term s b) -> Term s b

Pattern match over Plutarch Terms via a Haskell datatype

pcon :: forall a (s :: S). PlutusType a => a s -> Term s a

Construct a Plutarch Term via a Haskell datatype

type family PInner (a :: S -> Type) :: S -> Type

class PlutusTypeStrat strategy

Deprecated: Use the new mechanisms instead

Minimal complete definition

derivedPCon, derivedPMatch

class (PInner a ~ DerivedPInner (DPTStrat a) a, PlutusTypeStrat (DPTStrat a), PlutusTypeStratConstraint (DPTStrat a) a, PlutusType a) => DerivePlutusType (a :: S -> Type)

Deprecated: Use the new mechanisms instead

type family DPTStrat (a :: S -> Type)

type family DerivedPInner strategy (a :: S -> Type) :: S -> Type

derivedPCon :: forall a (s :: S). (PlutusTypeStrat strategy, DerivePlutusType a, DPTStrat a ~ strategy) => a s -> Term s (DerivedPInner strategy a)

derivedPMatch :: forall a (s :: S) (b :: S -> Type). (PlutusTypeStrat strategy, DerivePlutusType a, DPTStrat a ~ strategy) => Term s (DerivedPInner strategy a) -> (a s -> Term s b) -> Term s b

newtype DeriveNewtypePlutusType (a :: S -> Type) (s :: S)

Since: 1.10.0

Constructors

DeriveNewtypePlutusType

Fields

unDeriveNewtypePlutusType :: a s

Since: 1.10.0

newtype DeriveFakePlutusType (a :: S -> Type) (s :: S)

This is a cursed derivation strategy that will give you PlutusType with no questions asked. This is occasionally helpful for deriving PlutusType for another derivation strategy wrapper whose target instance requires PlutusType as superclass.

See PLiftable

Constructors

DeriveFakePlutusType (a s)