Plutarch.Internal.Term

Documentation

$hoisted

newtype ((a :: S -> Type) :--> (b :: S -> Type)) (s :: S) infixr 0

Constructors

PLam (Term s a -> Term s b)

data PDelayed (a :: S -> Type) (s :: S)

$term

newtype Term (s :: S) (a :: S -> Type)

Constructors

Term

Fields

asRawTerm :: ExceptT Text (RWS TermEnv () ()) TermResult

newtype Script

Constructors

Script (Program DeBruijn DefaultUni DefaultFun ())

mapTerm :: (RawTerm -> RawTerm) -> TermResult -> TermResult

plam' :: forall (s :: S) (a :: S -> Type) (b :: S -> Type). (Term s a -> Term s b) -> Term s (a :--> b)

Lambda abstraction.

Only works with a single argument. Use plam instead, to support currying.

plet :: forall (s :: S) (a :: S -> Type) (b :: S -> Type). Term s a -> (Term s a -> Term s b) -> Term s b

Let bindings.

This is approximately a shorthand for a lambda and application:

plet v f == papp (plam f) v

But sufficiently small terms in WHNF may be inlined for efficiency.

papp :: forall (s :: S) (a :: S -> Type) (b :: S -> Type). Term s (a :--> b) -> Term s a -> Term s b

Lambda Application.

pdelay :: forall (s :: S) (a :: S -> Type). Term s a -> Term s (PDelayed a)

Plutus 'delay', used for laziness.

pforce :: forall (s :: S) (a :: S -> Type). Term s (PDelayed a) -> Term s a

Plutus 'force', used to force evaluation of PDelayed terms.

phoistAcyclic :: forall (a :: S -> Type) (s :: S). HasCallStack => (forall (s' :: S). Term s' a) -> Term s a

perror :: forall (s :: S) (a :: S -> Type). Term s a

Plutus 'error'.

When using this explicitly, it should be ensured that the containing term is delayed, avoiding premature evaluation.

pplaceholder :: forall (s :: S) (a :: S -> Type). Integer -> Term s a

Same as perror except this holds integer id for AST look-ahead.

This can be used to "tag" branch and generate AST first to see if that branch is actually used or not, allowing optimization cutting unused branches. For more detailed uscases, check pmatchDataRec.

punsafeCoerce :: forall (b :: S -> Type) (a :: S -> Type) (s :: S). Term s a -> Term s b

Unsafely coerce the type-tag of a Term.

This should mostly be avoided, though it can be safely used to assert known types of Datums, Redeemers or ScriptContext.

punsafeBuiltin :: forall (s :: S) (a :: S -> Type). DefaultFun -> Term s a

punsafeConstantInternal :: forall (s :: S) (a :: S -> Type). Some (ValueOf DefaultUni) -> Term s a

compile :: forall (a :: S -> Type). Config -> (forall (s :: S). Term s a) -> Either Text Script

Compile a (closed) Plutus Term to a usable script

compileWithInternalConfig :: forall (a :: S -> Type). InternalConfig -> Config -> (forall (s :: S). Term s a) -> Either Text Script

As compile but exposes InternalConfig options.

Since: 1.12.0

compileOptimized :: forall (a :: S -> Type). (forall (s :: S). Term s a) -> Either Text Script

As compile, but performs UPLC optimizations. Furthermore, this will always elide tracing (as if with NoTracing).

Since: 1.10.0

compileOptimizedWithInternalConfig :: forall (a :: S -> Type). InternalConfig -> (forall (s :: S). Term s a) -> Either Text Script

As compileOptimized but exposes InternalConfig options.

Since: 1.12.0

compile' :: TermResult -> Term DeBruijn DefaultUni DefaultFun ()

optimizeTerm :: forall (a :: S -> Type). (forall (s :: S). Term s a) -> forall (s :: S). Term s a

Given a closed Term, run the UPLC optimizer on it.

Important note

If the input term has any hoisted dependencies, these are completely erased by this process. Thus, if the resulting Term is used as part of a larger computation with the same dependencies, the Plutarch compiler will not be aware of them, and will not be able to optimize them properly. More concretely, in a case like this:

@ padd # optimizeTerm (f # pexpensive) # optimizeTerm (g # pexpensive) @

pexpensive will end up being duplicated entirely into each 'arm' of the padd, even though under normal circumstances it could be shared.

Thus, if you plan to use this, ensure that it is done 'as late as possible'; embedding Terms produced by optimizeTerm into larger computations can lead to size blowout if not done carefully.

data RawTerm

Constructors

RVar Word64

RLamAbs Word64 RawTerm

RApply RawTerm [RawTerm]

RForce RawTerm

RDelay RawTerm

RConstant (Some (ValueOf DefaultUni))

RBuiltin DefaultFun

RCompiled (Term DeBruijn DefaultUni DefaultFun ())

RError

RHoisted HoistedTerm

RPlaceHolder Integer

RConstr Word64 [RawTerm]

RCase RawTerm [RawTerm]

RLet RawTerm RawTerm

data HoistedTerm

Constructors

HoistedTerm

Fields

htHash :: Int

htRawTerm :: RawTerm

data TermResult

Constructors

TermResult

Fields

getTerm :: RawTerm

getDeps :: [HoistedTerm]

data S

Constructors

SI

pthrow :: forall (s :: S) (a :: S -> Type). HasCallStack => Text -> Term s a

data Config where

Configuration for Plutarch scripts at compile time. This indicates whether we want to trace, and if so, under what log level and mode.

Since: 1.6.0

Bundled Patterns

pattern NoTracing :: Config
Pattern for the config that does no tracing (also the default).

Since: 1.6.0

pattern Tracing :: LogLevel -> TracingMode -> Config
Pattern for a tracing config, with both its log level and mode.

Since: 1.6.0

data InternalConfig

Constructors

InternalConfig

Fields

internalConfig'dataRecPMatchOptimization :: Bool

internalConfig'phoistAcyclicEvalCheck :: Bool

data TracingMode

How to trace.

Since: 1.6.0

Constructors

DetTracing

DoTracing

DoTracingAndBinds

data LogLevel

What logging level we want to use.

Since: 1.6.0

Constructors

LogInfo

LogDebug

tracingMode :: Config -> Maybe TracingMode

If the config indicates that we want to trace, get its mode.

Since: 1.6.0

logLevel :: Config -> Maybe LogLevel

If the config indicates that we want to trace, get its log level.

Since: 1.6.0

pgetConfig :: forall (s :: S) (a :: S -> Type). (Config -> Term s a) -> Term s a

pgetInternalConfig :: forall (s :: S) (a :: S -> Type). (InternalConfig -> Term s a) -> Term s a

pwithInternalConfig :: forall (s :: S) (a :: S -> Type). InternalConfig -> Term s a -> Term s a

(#) :: forall (s :: S) (a :: S -> Type) (b :: S -> Type). Term s (a :--> b) -> Term s a -> Term s b infixl 8

High precedence infixl synonym of papp, to be used like function juxtaposition. e.g.:

>>> f # x # y
f x y

(#$) :: forall (s :: S) (a :: S -> Type) (b :: S -> Type). Term s (a :--> b) -> Term s a -> Term s b infixr 0

Low precedence infixr synonym of papp, to be used like $, in combination with #. e.g.:

>>> f # x #$ g # y # z
f x (g y z)