Plutarch.LedgerApi.AssocMap

Description

This module is designed to be imported qualified, as many of its identifiers clash with the Plutarch prelude.

Types

data PSortedMap (k :: S -> Type) (v :: S -> Type) (s :: S)

newtype PUnsortedMap (k :: S -> Type) (v :: S -> Type) (s :: S)

Since: 3.5.0

Constructors

PUnsortedMap (Term s (PAssocMap k v))

newtype PAssocMap (k :: S -> Type) (v :: S -> Type) (s :: S)

Since: 3.5.0

Constructors

PAssocMap (Term s (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))))

data MergeHandler (s :: S) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (c :: S -> Type)

Constructors

MergeHandler

Fields

mhBothPresent :: BothPresentHandler s k a b c

mhLeftPresent :: OnePresentHandler s k a c

mhRightPresent :: OnePresentHandler s k b c

data BothPresentHandler (s :: S) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (c :: S -> Type)

Constructors

DropBoth

HandleBoth (Term s (k :--> (a :--> (b :--> c))))

HandleOrDropBoth (Term s (k :--> (a :--> (b :--> PMaybe c))))

data OnePresentHandler (s :: S) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type)

Constructors

DropOne

HandleOne (Term s (k :--> (a :--> b)))

HandleOrDropOne (Term s (k :--> (a :--> PMaybe b)))

Functions

Creation

pempty :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (v :: S -> Type) (s :: S). PInner (t k v) ~ PAssocMap k v => Term s (t k v)

Construct an empty map.

Since: 2.0.0

psingleton :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (v :: S -> Type) (s :: S). (PInner (t k v) ~ PAssocMap k v, PIsData k, PIsData v) => Term s (k :--> (v :--> t k v))

Construct a singleton map with the given key and value.

Since: 2.1.1

psingletonData :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (v :: S -> Type) (s :: S). PInner (t k v) ~ PAssocMap k v => Term s (PAsData k :--> (PAsData v :--> t k v))

Construct a singleton map with the given data-encoded key and value.

Since: 2.1.1

punsortedMapFromFoldable :: forall (k :: S -> Type) (v :: S -> Type) f (s :: S). (Foldable f, PIsData k, PIsData v) => f (Term s k, Term s v) -> Term s (PUnsortedMap k v)

Since: 2.1.1

psortedMapFromFoldable :: forall (k :: S -> Type) (v :: S -> Type) f (s :: S). (Foldable f, POrd k, PIsData k, PIsData v) => f (Term s k, Term s v) -> Term s (PSortedMap k v)

Since: 2.1.1

Transformation

passertSorted :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k) => Term s (PUnsortedMap k v :--> PSortedMap k v)

Deprecated: Use ppromoteToSortedMap instead

Attempt to promote PUnsortedMap to PSortedMap. This function checks that the keys in the input map are in strictly ascending order and fails with an error if they are not. Duplicate keys are not allowed.

Since: 2.0.0

ppromoteToSortedMap :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k) => Term s (PUnsortedMap k v :--> PSortedMap k v)

Attempt to promote PUnsortedMap to PSortedMap. This function checks that the keys in the input map are in strictly ascending order and fails with an error if they are not. Duplicate keys are not allowed.

Since: 3.6.0

punsafeCoerceToSortedMap :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). Term s (PUnsortedMap k v) -> Term s (PSortedMap k v)

Coerce PUnsortedMap to PSortedMap. Unsafe.

Since: 3.6.0

pforgetSorted :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). Term s (PSortedMap k v) -> Term s (PUnsortedMap k v)

Forget the knowledge that keys were sorted.

Since: 2.1.1

pmap :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PInner (t k a) ~ PAssocMap k a, PInner (t k b) ~ PAssocMap k b, PIsData a, PIsData b) => Term s ((a :--> b) :--> (t k a :--> t k b))

Applies a function to every value in the map, much like map.

Since: 2.0.0

pmapData :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PInner (t k a) ~ PAssocMap k a, PInner (t k b) ~ PAssocMap k b) => Term s ((PAsData a :--> PAsData b) :--> (t k a :--> t k b))

As pmap, but over Data representations.

Since: 2.0.0

pmapWithKey :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PInner (t k a) ~ PAssocMap k a, PInner (t k b) ~ PAssocMap k b, PIsData k, PIsData a, PIsData b) => Term s ((k :--> (a :--> b)) :--> (t k a :--> t k b))

As pmap, but gives key access as well.

Since: 2.1.1

pmapMaybe :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PInner (t k a) ~ PAssocMap k a, PInner (t k b) ~ PAssocMap k b, PIsData a, PIsData b) => Term s ((a :--> PMaybe b) :--> (t k a :--> t k b))

Maps and filters the map, much like mapMaybe.

Since: 2.0.0

pmapMaybeData :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PInner (t k a) ~ PAssocMap k a, PInner (t k b) ~ PAssocMap k b) => Term s ((PAsData a :--> PMaybe (PAsData b)) :--> (t k a :--> t k b))

As pmapMaybe, but over Data representation.

Since: 2.0.0

pmapMaybeWithKey :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PInner (t k a) ~ PAssocMap k a, PInner (t k b) ~ PAssocMap k b, PIsData k, PIsData a, PIsData b) => Term s ((k :--> (a :--> PMaybe b)) :--> (t k a :--> t k b))

As pmapMaybe, but gives key access as well.

Since: 3.5.0

pmapMaybeDataWithKey :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PInner (t k a) ~ PAssocMap k a, PInner (t k b) ~ PAssocMap k b) => Term s ((PAsData k :--> (PAsData a :--> PMaybe (PAsData b))) :--> (t k a :--> t k b))

As pmapMaybeData, but gives key access as well.

Since: 3.5.0

Relational lift

pcheckBinRel :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData v) => Term s ((v :--> (v :--> PBool)) :--> (v :--> (PSortedMap k v :--> (PSortedMap k v :--> PBool))))

Given a comparison function and a "zero" value, check whether a binary relation holds over two PSortedMaps.

Important note

This is primarily intended to be used with PValue. We assume that the comparison behaves like a comparison would (thus, being at least a partial order, or possibly a total order or equivalence), and that the starting value does not break it. Use with extreme care.

Since: 2.0.0

Comparison

pkeysEqual :: forall (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PEq k, PIsData k) => Term s (PSortedMap k a :--> (PSortedMap k b :--> PBool))

Gives PTrue if both argument PSortedMaps contain mappings for exactly the same set of keys. Requires a number of equality comparisons between keys proportional to the length of the shorter argument.

Since: 2.1.1

pkeysEqualUnsorted :: forall (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PIsData k, PIsData a, PIsData b) => Term s (PUnsortedMap k a :--> (PUnsortedMap k b :--> PBool))

As pkeysEqual, but requires only PEq constraints for the keys, and works for PUnsortedMaps. This requires a number of equality comparisons between keys proportional to the product of the lengths of both arguments: that is, this function is quadratic.

NOTE

This function does not handle duplicate keys in unsorted maps. For example, if one map contains duplicate keys while the other does not, the function may still return PTrue as long as each key appears at least once in both maps.

Use with caution - or preferably avoid using this function entirely - as it may be deprecated in a future release due to its non-obvious behavior.

Since: 2.1.1

Fold

pall :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). PIsData v => Term s ((v :--> PBool) :--> (PUnsortedMap k v :--> PBool))

Verifies all values in the map satisfy the given predicate.

Since: 2.0.0

pany :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). PIsData v => Term s ((v :--> PBool) :--> (PUnsortedMap k v :--> PBool))

Tests if any value in the map satisfies the given predicate.

Since: 2.1.1

pfoldMapWithKey :: forall (m :: S -> Type) (k :: S -> Type) (v :: S -> Type) (s :: S). (PIsData k, PIsData v, forall (s' :: S). Monoid (Term s' m)) => Term s ((k :--> (v :--> m)) :--> (PUnsortedMap k v :--> m))

Project all key-value pairs into a Monoid, then combine. Keys and values will be presented in key order.

Since: 3.6.0

pfoldlWithKey :: forall (a :: S -> Type) (k :: S -> Type) (v :: S -> Type) (s :: S). (PIsData k, PIsData v) => Term s ((a :--> (k :--> (v :--> a))) :--> (a :--> (PUnsortedMap k v :--> a)))

Left-associative fold with keys. Keys and values will be presented in key order.

Since: 3.6.0

Combination

zipWithBuilder :: forall (s :: S) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (c :: S -> Type). (POrd k, PIsData k, PIsData a, PIsData b, PIsData c) => MergeHandler s k a b c -> Term s (PSortedMap k a :--> (PSortedMap k b :--> PSortedMap k c))

Build a function that zips two PSortedMaps together using a custom MergeHandler.

The provided MergeHandler determines how to merge entries based on whether a key is present in the left map, the right map, or both.

NOTE

This function itself cannot be hoisted with phoistAcyclic because it depends on the supplied MergeHandler. However, once you specialize it by providing a specific merge handler, the resulting function should be hoisted if it will be reused, to avoid duplication.

Since: 3.5.0

zipWithDataBuilder :: forall (s :: S) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (c :: S -> Type). (POrd k, PIsData k) => MergeHandler s (PAsData k) (PAsData a) (PAsData b) (PAsData c) -> Term s (PSortedMap k a :--> (PSortedMap k b :--> PSortedMap k c))

Build a function that zips two PSortedMaps together using a custom MergeHandler.

The provided MergeHandler determines how to merge entries based on whether a key is present in the left map, the right map, or both.

Unlike zipWithBuilder, zipWithDataBuilder operates on values wrapped in PAsData (typed BuiltinData).

NOTE

This function itself cannot be hoisted with phoistAcyclic because it depends on the supplied MergeHandler. However, once you specialize it by providing a specific merge handler, the resulting function should be hoisted if it will be reused, to avoid duplication.

Since: 3.5.0

punionWith :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData v) => Term s ((v :--> (v :--> v)) :--> (PSortedMap k v :--> (PSortedMap k v :--> PSortedMap k v)))

Build the union of two PSortedMaps, merging values that share the same key using the given function.

Since: 3.5.0

punionWithData :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k) => Term s ((PAsData v :--> (PAsData v :--> PAsData v)) :--> (PSortedMap k v :--> (PSortedMap k v :--> PSortedMap k v)))

Build the union of two PSortedMaps, merging values that share the same key using the given function.

Since: 3.5.0

pleftBiasedUnion :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData v) => Term s (PSortedMap k v :--> (PSortedMap k v :--> PSortedMap k v))

Build the union of two PSortedMaps. Take the value from the left argument for colliding keys.

Since: 2.1.1

pdifference :: forall (b :: S -> Type) (a :: S -> Type) (k :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData a, PIsData b) => Term s (PSortedMap k a :--> (PSortedMap k b :--> PSortedMap k a))

pdifferenceWith :: forall (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData a, PIsData b) => Term s ((a :--> (b :--> PMaybe a)) :--> (PSortedMap k a :--> (PSortedMap k b :--> PSortedMap k a)))

pzipWithDefaults :: forall (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (c :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData a, PIsData b, PIsData c) => (forall (s' :: S). Term s' a) -> (forall (s' :: S). Term s' b) -> Term s ((a :--> (b :--> c)) :--> (PSortedMap k a :--> (PSortedMap k b :--> PSortedMap k c)))

Zip two PSortedMaps, using the given value merge function for key collisions, and different values for the sides.

Since: 2.1.1

pintersectionWith :: forall (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (c :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData a, PIsData b, PIsData c) => Term s ((a :--> (b :--> c)) :--> (PSortedMap k a :--> (PSortedMap k b :--> PSortedMap k c)))

Build the intersection of two PSortedMaps, merging values that share the same key using the given function.

Since: 2.1.1

pintersectionWithData :: forall (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (c :: S -> Type) (s :: S). (POrd k, PIsData k) => Term s ((PAsData a :--> (PAsData b :--> PAsData c)) :--> (PSortedMap k a :--> (PSortedMap k b :--> PSortedMap k c)))

Build the intersection of two PSortedMaps, merging data-encoded values that share the same key using the given function.

Since: 2.1.1

Query

pnull :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). Term s (PUnsortedMap k v :--> PBool)

Tests whether the map is empty.

Since: 2.0.0

plookup :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (PIsData k, PIsData v) => Term s (k :--> (PSortedMap k v :--> PMaybe v))

Look up the given key in a PSortedMap.

Since: 3.6.0

plookupData :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). Term s (PAsData k :--> (PSortedMap k v :--> PMaybe (PAsData v)))

As plookup, except over Data representation.

Since: 3.6.0

plookupDataWith :: forall (k :: S -> Type) (v :: S -> Type) (x :: S -> Type) (s :: S). Term s ((PBuiltinPair (PAsData k) (PAsData v) :--> PMaybe x) :--> (PAsData k :--> (PSortedMap k v :--> PMaybe x)))

Look up the given key data in a PSortedMap, applying the given function to the found key-value pair.

Since: 3.6.0

pfindWithDefault :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (PIsData k, PIsData v) => Term s (v :--> (k :--> (PSortedMap k v :--> v)))

Look up the given key in a PSortedMap, returning the default value if the key is absent.

Since: 3.6.0

pfoldAt :: forall (k :: S -> Type) (v :: S -> Type) (r :: S -> Type) (s :: S). PIsData k => Term s (k :--> (r :--> ((PAsData v :--> r) :--> (PSortedMap k v :--> r))))

Look up the given key in a PSortedMap; return the default if the key is absent or apply the argument function to the value data if present.

Since: 3.6.0

pfoldAtData :: forall (k :: S -> Type) (v :: S -> Type) (r :: S -> Type) (s :: S). Term s (PAsData k :--> (r :--> ((PAsData v :--> r) :--> (PSortedMap k v :--> r))))

Look up the given key data in a PSortedMap; return the default if the key is absent or apply the argument function to the value data if present.

Since: 3.6.0

ptryLookup :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (PIsData k, PIsData v) => Term s (k :--> (PSortedMap k v :--> v))

As plookup, but errors when the key is missing.

Since: 3.6.0

Modification

pinsert :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData v) => Term s (k :--> (v :--> (PSortedMap k v :--> PSortedMap k v)))

Insert a new key/value pair into the map, overriding the previous if any.

Since: 2.1.1

pdelete :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k) => Term s (k :--> (PSortedMap k v :--> PSortedMap k v))

Delete a key from the map.

Since: 2.1.1

pupdate :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (PIsData k, PIsData v, POrd k) => Term s ((v :--> PMaybe v) :--> (k :--> (PSortedMap k v :--> PSortedMap k v)))

Given an 'updater' and a key, if the key exists in the PMap, apply the 'updater' to it, otherwise do nothing. If the 'updater' produces PNothing, the value is deleted; otherwise, it is modified to the result.

Performance will be equivalent to a lookup followed by an insert (or delete), as well as the cost of calling the 'updater'.

Since: 2.1.1

padjust :: forall (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (v :: S -> Type) (s :: S). (PInner (t k v) ~ PAssocMap k v, PIsData k, PEq k, PIsData v) => Term s ((v :--> v) :--> (k :--> (t k v :--> t k v)))

If a value exists at the specified key, apply the function argument to it; otherwise, do nothing.

Since: 2.1.1

Key-value pair manipulation

pkvPairKey :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). PIsData k => Term s (PBuiltinPair (PAsData k) (PAsData v) :--> k)

Get the key of a key-value pair.

Since: 2.1.1

pkvPairValue :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). PIsData v => Term s (PBuiltinPair (PAsData k) (PAsData v) :--> v)

Get the value of a key-value pair.

Since: 2.1.1

pkvPairLt :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (PIsData k, POrd k) => Term s (PBuiltinPair (PAsData k) (PAsData v) :--> (PBuiltinPair (PAsData k) (PAsData v) :--> PBool))

Compare two key-value pairs by their keys. Gives PTrue if the key of the first argument pair is less than the key of the second argument pair.

Since: 2.1.1

Conversion

pkeys :: forall (ell :: (S -> Type) -> S -> Type) (t :: (S -> Type) -> (S -> Type) -> S -> Type) (k :: S -> Type) (v :: S -> Type) (s :: S). (PInner (t k v) ~ PAssocMap k v, PListLike ell, PElemConstraint ell (PAsData k)) => Term s (t k v :--> ell (PAsData k))

Extract the keys from the given PIsAssocMap instance as a list-like structure. If the provided Map is PSortedMap, the keys will maintain that order, and will be unique; otherwise, the order is unspecified, and duplicates may exist.

Note

You will need to specify what manner of list-like structure you want; we have arranged the type signature to make specifying this easy with TypeApplications.

Since: 2.1.1