Plutarch.LedgerApi.V3.Contexts

Documentation

newtype PColdCommitteeCredential (s :: S)

Since: 3.1.0

Constructors

PColdCommitteeCredential (Term s PCredential)

newtype PHotCommitteeCredential (s :: S)

Since: 3.1.0

Constructors

PHotCommitteeCredential (Term s PCredential)

newtype PDRepCredential (s :: S)

Since: 3.1.0

Constructors

PDRepCredential (Term s PCredential)

data PDRep (s :: S)

Since: 3.1.0

Constructors

PDRep (Term s (PAsData PDRepCredential))

PDRepAlwaysAbstain

PDRepAlwaysNoConfidence

data PDelegatee (s :: S)

Since: 3.1.0

Constructors

PDelegStake (Term s (PAsData PPubKeyHash))

PDelegVote (Term s PDRep)

PDelegStakeVote (Term s (PAsData PPubKeyHash)) (Term s PDRep)

data PTxCert (s :: S)

Since: 3.1.0

Constructors

PTxCertRegStaking (Term s PCredential) (Term s (PMaybeData PLovelace))

PTxCertUnRegStaking (Term s PCredential) (Term s (PMaybeData PLovelace))

PTxCertDelegStaking (Term s PCredential) (Term s PDelegatee)

PTxCertRegDeleg (Term s PCredential) (Term s PDelegatee) (Term s (PAsData PLovelace))

PTxCertRegDRep (Term s PDRepCredential) (Term s (PAsData PLovelace))

PTxCertUpdateDRep (Term s PDRepCredential)

PTxCertUnRegDRep (Term s PDRepCredential) (Term s (PAsData PLovelace))

PTxCertPoolRegister (Term s (PAsData PPubKeyHash)) (Term s (PAsData PPubKeyHash))

PTxCertPoolRetire (Term s (PAsData PPubKeyHash)) (Term s (PAsData PInteger))

PTxCertAuthHotCommittee (Term s PColdCommitteeCredential) (Term s PHotCommitteeCredential)

PTxCertResignColdCommittee (Term s PColdCommitteeCredential)

data PVoter (s :: S)

Since: 3.1.0

Constructors

PCommitteeVoter (Term s PHotCommitteeCredential)

PDRepVoter (Term s PDRepCredential)

PStakePoolVoter (Term s (PAsData PPubKeyHash))

data PVote (s :: S)

Since: 3.1.0

Constructors

PVoteYes

PVoteNo

PAbstain

data PGovernanceActionId (s :: S)

Since: 3.1.0

Constructors

PGovernanceActionId (Term s (PAsData PTxId)) (Term s (PAsData PInteger))

data PCommittee (s :: S)

Since: 3.1.0

Constructors

PCommittee

Fields

pcommittee'members :: Term s (PAsData (PUnsortedMap PColdCommitteeCredential PInteger))

pcommittee'quorum :: Term s PRationalData

newtype PConstitution (s :: S)

A constitution, omitting the optional anchor.

Since: 3.1.0

Constructors

PConstitution (Term s (PMaybeData PScriptHash))

data PProtocolVersion (s :: S)

Since: 3.1.0

Constructors

PProtocolVersion

Fields

pprotocolVersion'major :: Term s (PAsData PInteger)

pprotocolVersion'minor :: Term s (PAsData PInteger)

newtype PChangedParameters (s :: S)

Since: 3.1.0

Constructors

PChangedParameters (Term s PData)

data PGovernanceAction (s :: S)

Since: 3.1.0

Constructors

PParameterChange (Term s (PMaybeData PGovernanceActionId)) (Term s PChangedParameters) (Term s (PMaybeData PScriptHash))

PHardForkInitiation (Term s (PMaybeData PGovernanceActionId)) (Term s PProtocolVersion)

PTreasuryWithdrawals (Term s (PAsData (PUnsortedMap PCredential PLovelace))) (Term s (PMaybeData PScriptHash))

PNoConfidence (Term s (PMaybeData PGovernanceActionId))

PUpdateCommittee (Term s (PMaybeData PGovernanceActionId)) (Term s (PAsData (PBuiltinList (PAsData PColdCommitteeCredential)))) (Term s (PAsData (PUnsortedMap PColdCommitteeCredential PInteger))) (Term s PRationalData)

PNewConstitution (Term s (PMaybeData PGovernanceActionId)) (Term s PConstitution)

PInfoAction

data PProposalProcedure (s :: S)

Since: 3.1.0

Constructors

PProposalProcedure

Fields

pproposalProcedure'deposit :: Term s (PAsData PLovelace)

pproposalProcedure'returnAddr :: Term s PCredential

pproposalProcedure'governanceAction :: Term s PGovernanceAction

data PScriptPurpose (s :: S)

Since: 2.0.0

Constructors

PMinting (Term s (PAsData PCurrencySymbol))

PSpending (Term s PTxOutRef)

PRewarding (Term s PCredential)
Since: 3.1.0

PCertifying (Term s (PAsData PInteger)) (Term s PTxCert)

PVoting (Term s PVoter)
Since: 3.1.0

PProposing (Term s (PAsData PInteger)) (Term s PProposalProcedure)
Since: 3.1.0

data PScriptInfo (s :: S)

Since: 3.1.0

Constructors

PMintingScript (Term s (PAsData PCurrencySymbol))

PSpendingScript (Term s PTxOutRef) (Term s (PMaybeData PDatum))

PRewardingScript (Term s PCredential)

PCertifyingScript (Term s (PAsData PInteger)) (Term s PTxCert)

PVotingScript (Term s PVoter)

PProposingScript (Term s (PAsData PInteger)) (Term s PProposalProcedure)

data PTxInInfo (s :: S)

An input of the transaction.

Since: 2.0.0

Constructors

PTxInInfo

Fields

ptxInInfo'outRef :: Term s PTxOutRef

ptxInInfo'resolved :: Term s PTxOut

data PTxInfo (s :: S)

Constructors

PTxInfo

Fields

ptxInfo'inputs :: Term s (PAsData (PBuiltinList (PAsData PTxInInfo)))

ptxInfo'referenceInputs :: Term s (PAsData (PBuiltinList (PAsData PTxInInfo)))

ptxInfo'outputs :: Term s (PAsData (PBuiltinList (PAsData PTxOut)))

ptxInfo'fee :: Term s (PAsData PLovelace)

ptxInfo'mint :: Term s (PAsData PMintValue)

ptxInfo'txCerts :: Term s (PAsData (PBuiltinList (PAsData PTxCert)))

ptxInfo'wdrl :: Term s (PAsData (PUnsortedMap PCredential PLovelace))

ptxInfo'validRange :: Term s (PInterval PPosixTime)

ptxInfo'signatories :: Term s (PAsData (PBuiltinList (PAsData PPubKeyHash)))

ptxInfo'redeemers :: Term s (PAsData (PUnsortedMap PScriptPurpose PRedeemer))

ptxInfo'data :: Term s (PAsData (PUnsortedMap PDatumHash PDatum))

ptxInfo'id :: Term s (PAsData PTxId)

ptxInfo'votes :: Term s (PAsData (PUnsortedMap PVoter (PUnsortedMap PGovernanceActionId PVote)))

ptxInfo'proposalProcedures :: Term s (PAsData (PBuiltinList (PAsData PProposalProcedure)))

ptxInfo'currentTreasuryAmount :: Term s (PMaybeData PLovelace)

ptxInfo'treasuryDonation :: Term s (PMaybeData PLovelace)

data PScriptContext (s :: S)

Since: 3.1.0

Constructors

PScriptContext

Fields

pscriptContext'txInfo :: Term s PTxInfo

pscriptContext'redeemer :: Term s PRedeemer

pscriptContext'scriptInfo :: Term s PScriptInfo

pfindOwnInput :: forall (s :: S). Term s (PScriptContext :--> PMaybe PTxInInfo)

Find the input currently being validated.

Since: 3.6.0

pfindDatum :: forall (s :: S). Term s (PDatumHash :--> (PTxInfo :--> PMaybe PDatum))

Find the datum corresponding to a datum hash, if there is one.

Since: 3.1.0

pfindDatumHash :: forall (s :: S). Term s (PDatum :--> (PTxInfo :--> PMaybe PDatumHash))

Find the hash of a datum if it's part of the pending transaction's hashes.

Since: 3.1.0

pparseDatum :: forall (a :: S -> Type) (s :: S). PTryFrom PData (PAsData a) => Term s (PDatumHash :--> (PUnsortedMap PDatumHash PDatum :--> PMaybe (PAsData a)))

Lookup up the datum given the datum hash.

Takes as argument the datum assoc list from a PTxInfo. Validates the datum using PTryFrom.

Example:

 pparseDatum MyType # datumHash #$ pfield "datums" # txinfo
 

Since: 2.1.2

pfindTxInByTxOutRef :: forall (s :: S). Term s (PTxOutRef :--> (PTxInfo :--> PMaybe PTxInInfo))

Given a UTXO reference and a transaction (PTxInfo), resolve it to one of the transaction's inputs (PTxInInfo). If no matching input exists, the result is PNothing.

NOTE

This only searches the true transaction inputs and not the referenced transaction inputs.

Since: 3.6.0

pfindContinuingOutputs :: forall (s :: S). Term s (PScriptContext :--> PBuiltinList PInteger)

Find the indices of all the outputs that pay to the same script address we are currently spending from, if any.

Since: 3.6.0

pgetContinuingOutputs :: forall (s :: S). Term s (PScriptContext :--> PBuiltinList (PAsData PTxOut))

Get all the outputs that pay to the same script address we are currently spending from, if any.

Since: 3.6.0

ptxSignedBy :: forall (s :: S). Term s (PTxInfo :--> (PPubKeyHash :--> PBool))

Check if a transaction was signed by the given public key.

Since: 3.6.0

ppubKeyOutputsAt :: forall (s :: S). Term s (PPubKeyHash :--> (PTxInfo :--> PBuiltinList (PAsData PLedgerValue)))

Get the Values paid to a public key address by a pending transaction.

Since: 3.6.0

pvaluePaidTo :: forall (s :: S). Term s (PTxInfo :--> (PPubKeyHash :--> PLedgerValue))

Get the total value paid to a public key address by a pending transaction.

Since: 3.6.0

pvalueSpent :: forall (s :: S). Term s (PTxInfo :--> PLedgerValue)

pvalueProduced :: forall (s :: S). Term s (PTxInfo :--> PLedgerValue)

pownCurrencySymbol :: forall (s :: S). Term s (PScriptContext :--> PMaybe PCurrencySymbol)

Get the PCurrencySymbol of the current minting policy script.

Since: 3.6.0

pspendsOutput :: forall (s :: S). Term s (PTxInfo :--> (PTxId :--> (PInteger :--> PBool)))

Check if the pending transaction spends a specific transaction output (identified by the hash of a transaction and an index into that transactions' outputs)

Since: 3.6.0