Plutarch.LedgerApi.Value

Types

newtype PRawValue (s :: S)

Represents Values without any guarantees.

Values of this type may be unsorted, have duplicate keys, contain empty token maps, and include entries with zero token quantities.

Since: 3.5.0

Constructors

PRawValue (Term s (PUnsortedMap PCurrencySymbol (PUnsortedMap PTokenName PInteger)))

data PSortedValue (s :: S)

Represents sorted, well-formed Values without empty token maps.

Duplicate currency symbols or duplicate token names within the same token map are not allowed (since wip).

Compared to PRawValue, this type provides stronger guarantees, though PSortedValues may still contain entries with zero token quantities.

Since: 3.5.0

data PLedgerValue (s :: S)

Represents sorted, well-formed Values with a mandatory Ada entry.

Duplicate currency symbols or duplicate token names within the same token map are not allowed (since wip).

Like PSortedValue, but requires the presence of an Ada entry, which may have a zero quantity. Values of this type may still contain entries with zero token quantities.

Since: 3.5.0

newtype PCurrencySymbol (s :: S)

Since: 2.0.0

Constructors

PCurrencySymbol (Term s PByteString)

newtype PTokenName (s :: S)

Since: 2.0.0

Constructors

PTokenName (Term s PByteString)

newtype PLovelace (s :: S)

Since: 2.2.0

Constructors

PLovelace (Term s PInteger)

newtype PAssetClass (s :: S)

Since: 3.3.0

Constructors

PAssetClass (Term s (PBuiltinPair (PAsData PCurrencySymbol) (PAsData PTokenName)))

Functions

Creation

PRawValue

pemptyRawValue :: forall (s :: S). Term s PRawValue

Construct an empty PRawValue.

Since: 3.5.0

psingletonRawValue :: forall (s :: S). Term s (PCurrencySymbol :--> (PTokenName :--> (PInteger :--> PRawValue)))

Construct a singleton PRawValue containing only the given quantity of the given currency.

Since: 3.5.0

psingletonRawValueData :: forall (s :: S). Term s (PAsData PCurrencySymbol :--> (PAsData PTokenName :--> (PAsData PInteger :--> PRawValue)))

Construct a singleton PRawValue containing only the given quantity of the given currency, taking data-encoded parameters.

Since: 3.5.0

PSortedValue

pemptySortedValue :: forall (s :: S). Term s PSortedValue

Construct an empty PSortedValue.

Since: 3.5.0

psingletonSortedValue :: forall (s :: S). Term s (PCurrencySymbol :--> (PTokenName :--> (PInteger :--> PSortedValue)))

Construct a singleton PSortedValue containing only the given quantity of the given currency.

Since: 3.5.0

psingletonSortedValueData :: forall (s :: S). Term s (PAsData PCurrencySymbol :--> (PAsData PTokenName :--> (PAsData PInteger :--> PSortedValue)))

Like psingletonSortedValue, but accepts data-encoded arguments.

Since: 3.5.0

PLedgerValue

pemptyLedgerValue :: forall (s :: S). Term s PLedgerValue

Construct an empty PLedgerValue with a mandatory zero Ada entry.

Since: 3.5.0

psingletonLedgerValue :: forall (s :: S). Term s (PCurrencySymbol :--> (PTokenName :--> (PInteger :--> PLedgerValue)))

Construct a singleton PLedgerValue containing the given quantity of the given currency, together with a mandatory Ada entry (which may be zero).

Since: 3.5.0

psingletonLedgerValueData :: forall (s :: S). Term s (PAsData PCurrencySymbol :--> (PAsData PTokenName :--> (PAsData PInteger :--> PLedgerValue)))

Like psingletonLedgerValue, but accepts data-encoded arguments.

Since: 3.5.0

PCurrencySymbol

padaSymbol :: forall (s :: S). Term s PCurrencySymbol

The PCurrencySymbol of the Ada currency.

Since: 2.1.1

padaSymbolData :: forall (s :: S). Term s (PAsData PCurrencySymbol)

Data-encoded PCurrencySymbol of the Ada currency.

Since: 2.1.1

PTokenName

padaToken :: forall (s :: S). Term s PTokenName

The PTokenName of the Ada currency.

Since: 2.1.1

Transformation

ppromoteToSortedValue :: forall (s :: S). Term s (PRawValue :--> PSortedValue)

Attempt to promote a PRawValue to PSortedValue.

The conversion succeeds only if the input Value is already sorted and does not contain empty token maps. Otherwise, the function fails with an error.

Since: 3.6.0

pforgetSorted :: forall (s :: S). Term s PSortedValue -> Term s PRawValue

Safely demote a PSortedValue to a PRawValue.

Since: 3.5.0

ptoLedgerValue :: forall (s :: S). Term s (PSortedValue :--> PLedgerValue)

Convert a PSortedValue to a PLedgerValue, inserting the mandatory Ada entry if missing.

Since: 3.5.0

Partial ordering

plt :: forall (s :: S). Term s PSortedValue -> Term s PSortedValue -> Term s PBool

Mimics the lt operation on plutus-ledger-api's Value.

Since: 3.5.0

pleq :: forall (s :: S). Term s PSortedValue -> Term s PSortedValue -> Term s PBool

Mimics the leq operation on plutus-ledger-api's Value.

Since: 3.5.0

pcheckBinRel :: forall (s :: S). Term s ((PInteger :--> (PInteger :--> PBool)) :--> (PSortedValue :--> (PSortedValue :--> PBool)))

Given a description of a relation on amounts, check whether that relation holds over PSortedValues.

Important note

This is intended for use with boolean comparison functions, which must define at least a partial order (total orders and equivalences are acceptable as well). Use of this with anything else is not guaranteed to give anything resembling a sensible answer. Use with extreme care.

Since: 2.0.0

Combination

pleftBiasedCurrencyUnion :: forall (s :: S). Term s (PSortedValue :--> (PSortedValue :--> PSortedValue))

Combine two PSortedValues, taking the tokens from the left only, if a currency occurs on both sides.

Since: 2.1.1

pleftBiasedTokenUnion :: forall (s :: S). Term s (PSortedValue :--> (PSortedValue :--> PSortedValue))

Combine two PSortedValues, taking the tokens from the left only, if a token name of the same currency occurs on both sides.

Since: 2.1.1

punionWith :: forall (s :: S). Term s ((PInteger :--> (PInteger :--> PInteger)) :--> (PSortedValue :--> (PSortedValue :--> PSortedValue)))

Combine two PSortedValues applying the given function to any pair of quantities with the same asset class.

Since: 3.5.0

punionWithData :: forall (s :: S). Term s ((PAsData PInteger :--> (PAsData PInteger :--> PAsData PInteger)) :--> (PSortedValue :--> (PSortedValue :--> PSortedValue)))

Combine two PSortedValues applying the given function to any pair of data-encoded quantities with the same asset class.

Since: 3.5.0

Queries

pvalueOf :: forall (s :: S). Term s (PSortedValue :--> (PCurrencySymbol :--> (PTokenName :--> PInteger)))

Get the quantity of the given currency in the PSortedValue.

Since: 3.6.0

plovelaceValueOf :: forall (s :: S). Term s (PSortedValue :--> PInteger)

Get the amount of Lovelace in the PSortedValue.

Since: 2.1.1

pisAdaOnlyValue :: forall (s :: S). Term s (PSortedValue :--> PBool)

Test if the PSortedValue contains nothing except an Ada entry.

Note

This function does not verify that Ada is positive and may return PTrue for zero or negative Ada amounts.

Since: 2.1.1

phasZeroTokenQuantities :: forall (s :: S). Term s (PRawValue :--> PBool)

Check if the given Value contains zero token quantities.

Since: 3.6.0

phasAdaEntry :: forall (s :: S). Term s (PSortedValue :--> PBool)

Check if the given PSortedValue contains an ADA entry (can be zero).

Since: 3.6.0

phasZeroAdaEntry :: forall (s :: S). Term s (PSortedValue :--> PBool)

Check if the given PSortedValue has a zero ADA entry.

Since: 3.6.0

Misc (internal use)

pinsertAdaEntry :: forall (s :: S). Term s (PSortedValue :--> PSortedValue)

Ensure that the given PSortedValue contains an Ada entry. If missing, a zero Ada entry is inserted at the head of the underlying sorted map.

pnormalizeNoAdaNonZeroTokens :: forall (s :: S). Term s (PSortedValue :--> PSortedValue)

Normalize the argument to contain no Ada entries and no zero token quantities.

pmapAmounts :: forall (s :: S). Term s ((PInteger :--> PInteger) :--> (PRawValue :--> PRawValue))