PlutusCore

Parser

parseProgram :: (MonadError ParserErrorBundle m, MonadQuote m) => Text -> m (Program TyName Name DefaultUni DefaultFun SrcSpan)

Parse a PLC program. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors. This passes "test" to the parser as the name of the input stream; to supply a name explicity, use `parse program name input`.

parseTerm :: (MonadError ParserErrorBundle m, MonadQuote m) => Text -> m (Term TyName Name DefaultUni DefaultFun SrcSpan)

Parse a PLC term. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseType :: (MonadError ParserErrorBundle m, MonadQuote m) => Text -> m (Type TyName DefaultUni SrcSpan)

Parse a PLC type. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

data SourcePos

data SrcSpan

The span between two source locations.

This corresponds roughly to the SrcSpan used by GHC, but we define our own version so we don't have to depend on ghc to use it.

The line and column numbers are 1-based, and the unit is Unicode code point (or Char).

Constructors

SrcSpan

Fields

srcSpanFile :: FilePath

srcSpanSLine :: Int

srcSpanSCol :: Int

srcSpanELine :: Int

srcSpanECol :: Int

Same as GHC's SrcSpan, srcSpanECol is usually one more than the column of the last character of the thing this SrcSpan is for (unless the last character is the line break).

data SrcSpans

Builtins

data Some (tag :: k -> Type) where

Bundled Patterns

pattern Some :: forall {k} tag (a :: k). () => tag a -> Some tag

data SomeTypeIn (uni :: Type -> Type)

A particular type from a universe.

Constructors

SomeTypeIn !(uni (Esc a))

data Kinded (uni :: Type -> Type) ta where

Constructors

Kinded :: forall k (uni :: Type -> Type) (a :: k). Typeable k => !(uni (Esc a)) -> Kinded uni (Esc a)

data ValueOf (uni :: Type -> Type) a

A value of a particular type from a universe.

Constructors

ValueOf !(uni (Esc a)) !a

someValueOf :: forall a uni. uni (Esc a) -> a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), given its explicit type tag.

someValue :: forall a (uni :: Type -> Type). Contains uni a => a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), provided its type is in the universe.

someValueType :: forall (uni :: Type -> Type). Some (ValueOf uni) -> SomeTypeIn uni

data Esc (a :: k)

"Escapes" a type of an arbitrary kind to fit into Type.

class Contains (uni :: Type -> Type) (a :: k) where

A class for enumerating types and fully instantiated type formers that uni contains. For example, a particular ExampleUni may have monomorphic types in it:

instance ExampleUni Contains Integer where ... instance ExampleUni Contains Bool where ...

as well as polymorphic ones:

instance ExampleUni Contains [] where ... instance ExampleUni Contains (,) where ...

as well as their instantiations:

instance ExampleUni Contains a => ExampleUni Contains [a] where ... instance (ExampleUni Contains a, ExampleUni Contains b) => ExampleUni Contains (a, b) where ...

(a universe can have any subset of the mentioned sorts of types, for example it's fine to have instantiated polymorphic types and not have uninstantiated ones and vice versa)

Note that when used as a constraint of a function Contains does not allow you to directly express things like "uni has the Integer, Bool and [] types and type formers", because [] is not fully instantiated. So you can only say "uni has Integer, Bool, [Integer], [Bool], [[Integer]], [[Bool]] etc" and such manual enumeration is annoying, so we'd really like to be able to say that uni has lists of arbitrary built-in types (including lists of lists etc). Contains does not allow that, but Includes does. For example, in the body of the following definition:

foo :: (uni Includes Integer, uni Includes Bool, uni Includes []) => ... foo = ...

you can make use of the fact that uni has lists of arbitrary included types (integers, booleans and lists).

Hence most of the time opt for using the more flexible Includes.

Includes is defined in terms of Contains, so you only need to provide a Contains instance per type from the universe and you'll get Includes for free.

Methods

knownUni :: uni (Esc a)

class Closed (uni :: Type -> Type) where

A universe is Closed, if it's known how to constrain every type from the universe and every type can be encoded to / decoded from a sequence of integer tags. The universe doesn't have to be finite and providing support for infinite universes is the reason why we encode a type as a sequence of integer tags as opposed to a single integer tag. For example, given

  data U a where
      UList :: !(U a) -> U [a]
      UInt  :: U Int

UList (UList UInt) can be encoded to [0,0,1] where 0 and 1 are the integer tags of the UList and UInt constructors, respectively.

Associated Types

type Everywhere (uni :: Type -> Type) (constr :: Type -> Constraint)

A constrant for "constr a holds for any a from uni".

Methods

encodeUni :: uni a -> [Int]

Encode a type as a sequence of Int tags. The opposite of decodeUni.

withDecodedUni :: (forall k (a :: k). Typeable k => uni (Esc a) -> DecodeUniM r) -> DecodeUniM r

Decode a type and feed it to the continuation.

bring :: Everywhere uni constr => proxy constr -> uni (Esc a) -> (constr a => r) -> r

Bring a constr a instance in scope, provided a is a type from the universe and constr holds for any type from the universe.

type family EverywhereAll (uni :: Type -> Type) (constrs :: [Type -> Constraint]) where ...

Equations

EverywhereAll uni ('[] :: [Type -> Constraint]) = ()

EverywhereAll uni (constr ': constrs) = (Everywhere uni constr, EverywhereAll uni constrs)

knownUniOf :: forall {k} uni (a :: k) proxy. Contains uni a => proxy a -> uni (Esc a)

Same as knownUni, but receives a proxy.

class GShow (t :: k -> Type) where

Methods

gshowsPrec :: forall (a :: k). Int -> t a -> ShowS

show :: Show a => a -> String

A specialised variant of showsPrec, using precedence context zero, and returning an ordinary String.

class GEq (f :: k -> Type) where

Methods

geq :: forall (a :: k) (b :: k). f a -> f b -> Maybe (a :~: b)

class HasUniApply (uni :: Type -> Type) where

A class for "uni has general type application".

Methods

uniApply :: forall k l (f :: k -> l) (a :: k). uni (Esc f) -> uni (Esc a) -> uni (Esc (f a))

Apply a type constructor to an argument.

matchUniApply

Arguments

:: uni tb
The type.

-> r
What to return if the type is not an application.

-> (forall k l (f :: k -> l) (a :: k). tb ~ Esc (f a) => uni (Esc f) -> uni (Esc a) -> r)
The continuation taking a function and an argument.

-> r

Deconstruct a type application into the function and the argument and feed them to the continuation. If the type is not an application, then return the default value.

checkStar :: forall uni a (x :: a). Typeable a => uni (Esc x) -> Maybe (a :~: Type)

Check if the kind of the given type from the universe is Type.

withApplicable :: forall a ab (f :: ab) (x :: a) uni m r. (Typeable ab, Typeable a, MonadPlus m) => uni (Esc f) -> uni (Esc x) -> (forall b. (Typeable b, ab ~ (a -> b)) => m r) -> m r

Check if one type from the universe can be applied to another (i.e. check that the expected kind of the argument matches the actual one) and call the continuation in the refined context. Fail with mzero otherwise.

data (a :: k) :~: (b :: k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: forall {k} (a :: k). a :~: a

type (<:) (uni1 :: Type -> Type) (uni2 :: Type -> Type) = Everywhere uni1 (Includes uni2 :: Type -> Constraint)

A constraint for "uni1 is a subuniverse of uni2".

type HasTypeLevel (uni :: Type -> Type) (x :: a) = KnownTypeAst Void uni (ElaborateBuiltin uni x)

Specifies that the given type is a built-in one and can be embedded into a Type.

type HasTermLevel (uni :: Type -> Type) = Includes uni :: a -> Constraint

Specifies that the given type is a built-in one and its values can be embedded into a Term.

type HasTypeAndTermLevel (uni :: Type -> Type) (x :: a) = (HasTypeLevel uni x, HasTermLevel uni x)

The product of HasTypeLevel and HasTermLevel.

data DefaultUni a where

The universe used by default.

Constructors

DefaultUniInteger :: DefaultUni (Esc Integer)

DefaultUniByteString :: DefaultUni (Esc ByteString)

DefaultUniString :: DefaultUni (Esc Text)

DefaultUniUnit :: DefaultUni (Esc ())

DefaultUniBool :: DefaultUni (Esc Bool)

DefaultUniProtoArray :: DefaultUni (Esc Vector)

DefaultUniProtoList :: DefaultUni (Esc [])

DefaultUniProtoPair :: DefaultUni (Esc (,))

DefaultUniApply :: forall {k1} {k2} (f :: k1 -> k2) (a1 :: k1). !(DefaultUni (Esc f)) -> !(DefaultUni (Esc a1)) -> DefaultUni (Esc (f a1))

DefaultUniData :: DefaultUni (Esc Data)

DefaultUniBLS12_381_G1_Element :: DefaultUni (Esc Element)

DefaultUniBLS12_381_G2_Element :: DefaultUni (Esc Element)

DefaultUniBLS12_381_MlResult :: DefaultUni (Esc MlResult)

DefaultUniValue :: DefaultUni (Esc Value)

pattern DefaultUniList :: forall {a} {k1} {k2} {f :: k1 -> k2} {a1 :: k1}. () => forall. (a ~ Esc (f a1), Esc f ~ Esc []) => DefaultUni (Esc a1) -> DefaultUni a

pattern DefaultUniPair :: forall {a} {k1} {k2} {f1 :: k1 -> k2} {a1 :: k1} {k3} {k4} {f2 :: k3 -> k4} {a2 :: k3}. () => forall. (a ~ Esc (f1 a1), Esc f1 ~ Esc (f2 a2), Esc f2 ~ Esc (,)) => DefaultUni (Esc a2) -> DefaultUni (Esc a1) -> DefaultUni a

pattern DefaultUniArray :: forall {a} {k1} {k2} {f :: k1 -> k2} {a1 :: k1}. () => forall. (a ~ Esc (f a1), Esc f ~ Esc Vector) => DefaultUni (Esc a1) -> DefaultUni a

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins/language versions and protocol versions]

Constructors

AddInteger

SubtractInteger

MultiplyInteger

DivideInteger

QuotientInteger

RemainderInteger

ModInteger

EqualsInteger

LessThanInteger

LessThanEqualsInteger

AppendByteString

ConsByteString

SliceByteString

LengthOfByteString

IndexByteString

EqualsByteString

LessThanByteString

LessThanEqualsByteString

Sha2_256

Sha3_256

Blake2b_256

VerifyEd25519Signature

VerifyEcdsaSecp256k1Signature

VerifySchnorrSecp256k1Signature

AppendString

EqualsString

EncodeUtf8

DecodeUtf8

IfThenElse

ChooseUnit

Trace

FstPair

SndPair

ChooseList

MkCons

HeadList

TailList

NullList

ChooseData

ConstrData

MapData

ListData

IData

BData

UnConstrData

UnMapData

UnListData

UnIData

UnBData

EqualsData

SerialiseData

MkPairData

MkNilData

MkNilPairData

Bls12_381_G1_add

Bls12_381_G1_neg

Bls12_381_G1_scalarMul

Bls12_381_G1_equal

Bls12_381_G1_hashToGroup

Bls12_381_G1_compress

Bls12_381_G1_uncompress

Bls12_381_G2_add

Bls12_381_G2_neg

Bls12_381_G2_scalarMul

Bls12_381_G2_equal

Bls12_381_G2_hashToGroup

Bls12_381_G2_compress

Bls12_381_G2_uncompress

Bls12_381_millerLoop

Bls12_381_mulMlResult

Bls12_381_finalVerify

Keccak_256

Blake2b_224

IntegerToByteString

ByteStringToInteger

AndByteString

OrByteString

XorByteString

ComplementByteString

ReadBit

WriteBits

ReplicateByte

ShiftByteString

RotateByteString

CountSetBits

FindFirstSetBit

Ripemd_160

ExpModInteger

DropList

LengthOfArray

ListToArray

IndexArray

Bls12_381_G1_multiScalarMul

Bls12_381_G2_multiScalarMul

InsertCoin

LookupCoin

UnionValue

ValueContains

ValueData

UnValueData

ScaleValue

AST

data Term tyname name (uni :: Type -> Type) fun ann

Constructors

Var ann name
a named variable

LamAbs ann name (Type tyname uni ann) (Term tyname name uni fun ann)
lambda abstraction

Apply ann (Term tyname name uni fun ann) (Term tyname name uni fun ann)
application

TyAbs ann tyname (Kind ann) (Term tyname name uni fun ann)
type abstraction

TyInst ann (Term tyname name uni fun ann) (Type tyname uni ann)
instantiation

IWrap ann (Type tyname uni ann) (Type tyname uni ann) (Term tyname name uni fun ann)
wrapping

Unwrap ann (Term tyname name uni fun ann)
unwrapping See Note [Constr tag type]

Constr ann (Type tyname uni ann) Word64 [Term tyname name uni fun ann]
constructor

Case ann (Type tyname uni ann) (Term tyname name uni fun ann) [Term tyname name uni fun ann]
case

Constant ann (Some (ValueOf uni))
constants

Builtin ann fun
builtin functions

Error ann (Type tyname uni ann)
fail with error

termSubterms :: forall tyname name (uni :: Type -> Type) fun ann f. Applicative f => (Term tyname name uni fun ann -> f (Term tyname name uni fun ann)) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the direct child Terms of the given Term.

termSubtypes :: forall tyname name (uni :: Type -> Type) fun ann f. Applicative f => (Type tyname uni ann -> f (Type tyname uni ann)) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the direct child Types of the given Term.

termMapNames :: forall tyname tyname' name name' (uni :: Type -> Type) fun ann. (tyname -> tyname') -> (name -> name') -> Term tyname name uni fun ann -> Term tyname' name' uni fun ann

programMapNames :: forall tyname tyname' name name' (uni :: Type -> Type) fun ann. (tyname -> tyname') -> (name -> name') -> Program tyname name uni fun ann -> Program tyname' name' uni fun ann

type family UniOf a :: Type -> Type

Extract the universe from a type.

data Type tyname (uni :: Type -> Type) ann

A Type assigned to expressions.

Constructors

TyVar ann tyname
Type variable

TyFun ann (Type tyname uni ann) (Type tyname uni ann)
Function type

TyIFix ann (Type tyname uni ann) (Type tyname uni ann)
Fix-point type, for constructing self-recursive types

TyForall ann tyname (Kind ann) (Type tyname uni ann)
Polymorphic type

TyBuiltin ann (SomeTypeIn uni)
Builtin type

TyLam ann tyname (Kind ann) (Type tyname uni ann)
Type lambda

TyApp ann (Type tyname uni ann) (Type tyname uni ann)
Type application

TySOP ann [[Type tyname uni ann]]
Sum-of-products type

typeSubtypes :: forall tyname (uni :: Type -> Type) ann f. Applicative f => (Type tyname uni ann -> f (Type tyname uni ann)) -> Type tyname uni ann -> f (Type tyname uni ann)

Get all the direct child Types of the given Type.

typeMapNames :: forall tyname tyname' (uni :: Type -> Type) ann. (tyname -> tyname') -> Type tyname uni ann -> Type tyname' uni ann

data Kind ann

Constructors

Type ann

KindArrow ann (Kind ann) (Kind ann)

toPatFuncKind :: Kind () -> Kind ()

The kind of a pattern functor (the first Type argument of TyIFix) at a given kind (of the second Type argument of TyIFix):

toPatFuncKind k = (k -> *) -> k -> * 

fromPatFuncKind :: Kind () -> Maybe (Kind ())

argsFunKind :: Kind ann -> [Kind ann]

Extract all a_i from a_0 -> a_1 -> ... -> r.

data ParserError

An error encountered during parsing.

Constructors

BuiltinTypeNotAStar !Text !SourcePos

UnknownBuiltinFunction !Text !SourcePos ![Text]

InvalidBuiltinConstant !Text !Text !SourcePos

data Version

The version of Plutus Core used by this program.

The intention is to convey different levels of backwards compatibility for existing scripts: - Major version changes are backwards-incompatible - Minor version changes are backwards-compatible - Patch version changes should be entirely invisible (and we will likely not use this level)

The version used should be changed only when the language itself changes. For example, adding a new kind of term to the language would require a minor version bump; removing a kind of term would require a major version bump.

Similarly, changing the semantics of the language will require a version bump, typically a major one. This is the main reason why the version is actually tracked in the AST: we can have two language versions with identical ASTs but different semantics, so we need to track the version explicitly.

Compatibility is about compatibility for specific scripts, not about e.g. tools which consume scripts. Adding a new kind of term does not change how existing scripts behave, but does change what tools would need to do to process scripts.

Constructors

Version

Fields

_versionMajor :: Natural

_versionMinor :: Natural

_versionPatch :: Natural

data Program tyname name (uni :: Type -> Type) fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn :: ann

_progVer :: Version

_progTerm :: Term tyname name uni fun ann

data Name

A Name represents variables/names in Plutus Core.

Constructors

Name

Fields

_nameText :: Text

The identifier name, for use in error messages.

_nameUnique :: Unique

A Unique assigned to the name, allowing for cheap comparisons in the compiler.

newtype TyName

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName :: Name

newtype Unique

A unique identifier This is normally a positive integral number, except in topUnique where we make use of a negative unique to signify top-level.

Constructors

Unique

Fields

unUnique :: Int

newtype UniqueMap unique a

A mapping from Uniques to arbitrary values of type a. Since Unique is equivalent to Int (see PlutusCore.Name.Unique), we can use an IntMap representation for this type.

Constructors

UniqueMap

Fields

unUniqueMap :: IntMap a

newtype UniqueSet unique

A set containing Uniques. Since Unique is equivalent to Int (see PlutusCore.Name.Unique), we can use an IntSet representation for this type.

Constructors

UniqueSet

Fields

unUniqueSet :: IntSet

newtype Normalized a

Constructors

Normalized

Fields

unNormalized :: a

latestVersion :: Version

The latest version of Plutus Core supported by this library.

class HasAnn (f :: Type -> Type) where

Types that have an outermost annotation.

Minimal complete definition

getAnn, modifyAnn

Methods

getAnn :: f a -> a

Get the outermost annotation.

modifyAnn :: (a -> a) -> f a -> f a

Modify the outermost annotation.

setAnn :: a -> f a -> f a

Set the outermost annotation.

tyVarDeclAnn :: forall tyname ann f. Functor f => (ann -> f ann) -> TyVarDecl tyname ann -> f (TyVarDecl tyname ann)

tyVarDeclName :: forall tyname1 ann tyname2 f. Functor f => (tyname1 -> f tyname2) -> TyVarDecl tyname1 ann -> f (TyVarDecl tyname2 ann)

tyVarDeclKind :: forall tyname ann f. Functor f => (Kind ann -> f (Kind ann)) -> TyVarDecl tyname ann -> f (TyVarDecl tyname ann)

varDeclAnn :: forall tyname name (uni :: Type -> Type) ann f. Functor f => (ann -> f ann) -> VarDecl tyname name uni ann -> f (VarDecl tyname name uni ann)

varDeclName :: forall tyname name1 (uni :: Type -> Type) ann name2 f. Functor f => (name1 -> f name2) -> VarDecl tyname name1 uni ann -> f (VarDecl tyname name2 uni ann)

varDeclType :: forall tyname1 name (uni1 :: Type -> Type) ann tyname2 (uni2 :: Type -> Type) f. Functor f => (Type tyname1 uni1 ann -> f (Type tyname2 uni2 ann)) -> VarDecl tyname1 name uni1 ann -> f (VarDecl tyname2 name uni2 ann)

tyDeclAnn :: forall tyname (uni :: Type -> Type) ann f. Functor f => (ann -> f ann) -> TyDecl tyname uni ann -> f (TyDecl tyname uni ann)

tyDeclType :: forall tyname1 (uni1 :: Type -> Type) ann tyname2 (uni2 :: Type -> Type) f. Functor f => (Type tyname1 uni1 ann -> f (Type tyname2 uni2 ann)) -> TyDecl tyname1 uni1 ann -> f (TyDecl tyname2 uni2 ann)

tyDeclKind :: forall tyname (uni :: Type -> Type) ann f. Functor f => (Kind ann -> f (Kind ann)) -> TyDecl tyname uni ann -> f (TyDecl tyname uni ann)

progAnn :: forall tyname name (uni :: Type -> Type) fun ann f. Functor f => (ann -> f ann) -> Program tyname name uni fun ann -> f (Program tyname name uni fun ann)

progVer :: forall tyname name (uni :: Type -> Type) fun ann f. Functor f => (Version -> f Version) -> Program tyname name uni fun ann -> f (Program tyname name uni fun ann)

progTerm :: forall tyname1 name1 (uni1 :: Type -> Type) fun1 ann tyname2 name2 (uni2 :: Type -> Type) fun2 f. Functor f => (Term tyname1 name1 uni1 fun1 ann -> f (Term tyname2 name2 uni2 fun2 ann)) -> Program tyname1 name1 uni1 fun1 ann -> f (Program tyname2 name2 uni2 fun2 ann)

mapFun :: forall fun fun' tyname name (uni :: Type -> Type) ann. (fun -> fun') -> Term tyname name uni fun ann -> Term tyname name uni fun' ann

Map a function over the set of built-in functions.

DeBruijn representation

newtype DeBruijn

A term name as a de Bruijn index, without the name string.

Constructors

DeBruijn

Fields

dbnIndex :: Index

newtype TyDeBruijn

A type name as a de Bruijn index, without the name string.

Constructors

TyDeBruijn DeBruijn

data NamedDeBruijn

A term name as a de Bruijn index.

Constructors

NamedDeBruijn

Fields

ndbnString :: !Text

ndbnIndex :: !Index

newtype NamedTyDeBruijn

A type name as a de Bruijn index.

Constructors

NamedTyDeBruijn NamedDeBruijn

deBruijnTerm :: forall m (uni :: Type -> Type) fun ann. MonadError FreeVariableError m => Term TyName Name uni fun ann -> m (Term NamedTyDeBruijn NamedDeBruijn uni fun ann)

Convert a Term with TyNames and Names into a Term with NamedTyDeBruijns and NamedDeBruijns. Will throw an error if a free variable is encountered.

unDeBruijnTerm :: forall m (uni :: Type -> Type) fun ann. (MonadQuote m, MonadError FreeVariableError m) => Term NamedTyDeBruijn NamedDeBruijn uni fun ann -> m (Term TyName Name uni fun ann)

Convert a Term with NamedTyDeBruijns and NamedDeBruijns into a Term with TyNames and Names. Will throw an error if a free variable is encountered.

Processing

type family HasUniques a

All kinds of uniques an entity contains.

class Rename a where

The class of things that can be renamed. I.e. things that are capable of satisfying the global uniqueness condition.

Methods

rename :: MonadQuote m => a -> m a

Rename Uniques so that they're globally unique. In case there are any free variables, they must be left untouched and bound variables must not get renamed to free ones. Must always assign new names to bound variables, so that rename can be used for alpha-renaming as well.

Type checking

module PlutusCore.TypeCheck

normalizeTypesIn :: forall tyname name (uni :: Type -> Type) m fun ann. (HasUnique tyname TypeUnique, HasUnique name TermUnique, MonadNormalizeType uni m) => Term tyname name uni fun ann -> m (Term tyname name uni fun ann)

Normalize every Type in a Term.

normalizeTypesInProgram :: forall tyname name (uni :: Type -> Type) m fun ann. (HasUnique tyname TypeUnique, HasUnique name TermUnique, MonadNormalizeType uni m) => Program tyname name uni fun ann -> m (Program tyname name uni fun ann)

Normalize every Type in a Program.

data TypeError term (uni :: Type -> Type) fun ann

Errors

data Error (uni :: Type -> Type) fun ann

Constructors

ParseErrorE !ParserErrorBundle

UniqueCoherencyErrorE !(UniqueError ann)

TypeErrorE !(TypeError (Term TyName Name uni fun ()) uni fun ann)

NormCheckErrorE !(NormCheckError TyName Name uni fun ann)

FreeVariableErrorE !FreeVariableError

data NormCheckError tyname name (uni :: Type -> Type) fun ann

Constructors

BadType !ann !(Type tyname uni ann) !Text

BadTerm !ann !(Term tyname name uni fun ann) !Text

data UniqueError ann

Constructors

MultiplyDefined !Unique !ann !ann

IncoherentUsage !Unique !ann !ann

FreeVariable !Unique !ann

data FreeVariableError

We cannot do a correct translation to or from de Bruijn indices if the program is not well-scoped. So we throw an error in such a case.

Constructors

FreeUnique !Unique

FreeIndex !Index

Quotation and term construction

type Quote = QuoteT Identity

A non-transformer version of QuoteT.

runQuote :: Quote a -> a

See runQuoteT.

data QuoteT (m :: Type -> Type) a

The "quotation" monad transformer. Within this monad you can do safe construction of PLC terms using quasiquotation, fresh-name generation, and parsing.

runQuoteT :: Monad m => QuoteT m a -> m a

Run a quote from an empty identifier state. Note that the resulting term cannot necessarily be safely combined with other terms - that should happen inside QuoteT.

class Monad m => MonadQuote (m :: Type -> Type)

A monad that allows lifting of quoted expressions.

liftQuote :: MonadQuote m => Quote a -> m a

Name generation

freshUnique :: MonadQuote m => m Unique

Get a fresh Unique.

freshName :: MonadQuote m => Text -> m Name

Get a fresh Name, given the annotation and the Text name.

freshTyName :: MonadQuote m => Text -> m TyName

Get a fresh TyName, given the annotation and the Text name.

Evaluation

data EvaluationResult a

The parameterized type of results various evaluation engines return.

Constructors

EvaluationSuccess !a

EvaluationFailure

Combining programs

applyProgram :: forall m a tyname name (uni :: Type -> Type) fun. (MonadError ApplyProgramError m, Semigroup a) => Program tyname name uni fun a -> Program tyname name uni fun a -> m (Program tyname name uni fun a)

Applies one program to another. Fails if the versions do not match and tries to merge annotations.

Benchmarking

termAstSize :: forall tyname name (uni :: Type -> Type) fun ann. Term tyname name uni fun ann -> AstSize

Count the number of AST nodes in a term.

typeAstSize :: forall tyname (uni :: Type -> Type) ann. Type tyname uni ann -> AstSize

Count the number of AST nodes in a type.

kindAstSize :: Kind a -> AstSize

Count the number of AST nodes in a kind.

>>> kindAstSize $ Type ()
AstSize {unAstSize = 1}
>>> kindAstSize $ KindArrow () (KindArrow () (Type ()) (Type ())) (Type ())
AstSize {unAstSize = 5} 

programAstSize :: forall tyname name (uni :: Type -> Type) fun ann. Program tyname name uni fun ann -> AstSize

Count the number of AST nodes in a program.