PlutusCore.Core.Plated

Documentation

kindSubkinds :: forall ann f. Applicative f => (Kind ann -> f (Kind ann)) -> Kind ann -> f (Kind ann)

kindSubkindsDeep :: forall ann f. (Contravariant f, Applicative f) => (Kind ann -> f (Kind ann)) -> Kind ann -> f (Kind ann)

tyVarDeclSubkinds :: forall tyname a f. Applicative f => (Kind a -> f (Kind a)) -> TyVarDecl tyname a -> f (TyVarDecl tyname a)

Get all the direct child Kinds of the given TyVarDecl.

typeTyBinds :: forall tyname (uni :: Type -> Type) ann f. Applicative f => (tyname -> f tyname) -> Type tyname uni ann -> f (Type tyname uni ann)

Get all the direct child 'tyname a's of the given Type from binders.

typeTyVars :: forall tyname (uni :: Type -> Type) ann f. Applicative f => (tyname -> f tyname) -> Type tyname uni ann -> f (Type tyname uni ann)

Get all the direct child 'tyname a's of the given Type from TyVars.

typeUniques :: forall tyname (uni :: Type -> Type) ann. HasUniques (Type tyname uni ann) => Traversal' (Type tyname uni ann) Unique

Get all the direct child Uniques of the given Type from binders TyVars.

typeSubkinds :: forall tyname (uni :: Type -> Type) ann f. Applicative f => (Kind ann -> f (Kind ann)) -> Type tyname uni ann -> f (Type tyname uni ann)

Get all the direct child Kinds of the given Type.

typeSubtypes :: forall tyname (uni :: Type -> Type) ann f. Applicative f => (Type tyname uni ann -> f (Type tyname uni ann)) -> Type tyname uni ann -> f (Type tyname uni ann)

Get all the direct child Types of the given Type.

typeSubtypesDeep :: forall tyname (uni :: Type -> Type) ann f. (Contravariant f, Applicative f) => (Type tyname uni ann -> f (Type tyname uni ann)) -> Type tyname uni ann -> f (Type tyname uni ann)

Get all the transitive child Types of the given Type.

varDeclSubtypes :: forall tyname name (uni :: Type -> Type) a f. Applicative f => (Type tyname uni a -> f (Type tyname uni a)) -> VarDecl tyname name uni a -> f (VarDecl tyname name uni a)

Get all the direct child Types of the given VarDecl.

termConstants :: forall tyname name (uni :: Type -> Type) fun ann f. Applicative f => (Some (ValueOf uni) -> f (Some (ValueOf uni))) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the direct constants of the given Term from Constants.

termTyBinds :: forall tyname name (uni :: Type -> Type) fun ann f. Applicative f => (tyname -> f tyname) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the direct child 'tyname a's of the given Term from TyAbses.

termBinds :: forall tyname name (uni :: Type -> Type) fun ann f. Applicative f => (name -> f name) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the direct child 'name a's of the given Term from LamAbses.

termVars :: forall tyname name (uni :: Type -> Type) fun ann f. Applicative f => (name -> f name) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the direct child 'name a's of the given Term from Vars.

termUniques :: forall tyname name (uni :: Type -> Type) fun ann. HasUniques (Term tyname name uni fun ann) => Traversal' (Term tyname name uni fun ann) Unique

Get all the direct child Uniques of the given Term (including the type-level ones).

termSubkinds :: forall tyname name (uni :: Type -> Type) fun ann f. Applicative f => (Kind ann -> f (Kind ann)) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the direct child Kinds of the given Term.

termSubtypes :: forall tyname name (uni :: Type -> Type) fun ann f. Applicative f => (Type tyname uni ann -> f (Type tyname uni ann)) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the direct child Types of the given Term.

termSubtermsDeep :: forall tyname name (uni :: Type -> Type) fun ann f. (Contravariant f, Applicative f) => (Term tyname name uni fun ann -> f (Term tyname name uni fun ann)) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the transitive child Terms of the given Term.

termSubtypesDeep :: forall tyname name (uni :: Type -> Type) fun ann f. (Contravariant f, Applicative f) => (Type tyname uni ann -> f (Type tyname uni ann)) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the transitive child Types of the given Term.

termConstantsDeep :: forall tyname name (uni :: Type -> Type) fun ann f. (Contravariant f, Applicative f) => (Some (ValueOf uni) -> f (Some (ValueOf uni))) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the transitive child Constants of the given Term.

termSubterms :: forall tyname name (uni :: Type -> Type) fun ann f. Applicative f => (Term tyname name uni fun ann -> f (Term tyname name uni fun ann)) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Get all the direct child Terms of the given Term.

typeUniquesDeep :: forall tyname (uni :: Type -> Type) ann. HasUniques (Type tyname uni ann) => Fold (Type tyname uni ann) Unique

Get all the transitive child Uniques of the given Type.

termUniquesDeep :: forall tyname name (uni :: Type -> Type) fun ann. HasUniques (Term tyname name uni fun ann) => Fold (Term tyname name uni fun ann) Unique

Get all the transitive child Uniques of the given Term (including the type-level ones).