PlutusCore.Default

Documentation

module PlutusCore.Default.Builtins

data (a :: k) :~: (b :: k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: forall {k} (a :: k). a :~: a

data Some (tag :: k -> Type) where

Bundled Patterns

pattern Some :: forall {k} tag (a :: k). () => tag a -> Some tag

data SomeTypeIn (uni :: Type -> Type)

A particular type from a universe.

Constructors

SomeTypeIn !(uni (Esc a))

data Kinded (uni :: Type -> Type) ta where

Constructors

Kinded :: forall k (uni :: Type -> Type) (a :: k). Typeable k => !(uni (Esc a)) -> Kinded uni (Esc a)

data ValueOf (uni :: Type -> Type) a

A value of a particular type from a universe.

Constructors

ValueOf !(uni (Esc a)) !a

someValueOf :: forall a uni. uni (Esc a) -> a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), given its explicit type tag.

someValue :: forall a (uni :: Type -> Type). Contains uni a => a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), provided its type is in the universe.

someValueType :: forall (uni :: Type -> Type). Some (ValueOf uni) -> SomeTypeIn uni

data Esc (a :: k)

"Escapes" a type of an arbitrary kind to fit into Type.

class Contains (uni :: Type -> Type) (a :: k) where

A class for enumerating types and fully instantiated type formers that uni contains. For example, a particular ExampleUni may have monomorphic types in it:

instance ExampleUni Contains Integer where ... instance ExampleUni Contains Bool where ...

as well as polymorphic ones:

instance ExampleUni Contains [] where ... instance ExampleUni Contains (,) where ...

as well as their instantiations:

instance ExampleUni Contains a => ExampleUni Contains [a] where ... instance (ExampleUni Contains a, ExampleUni Contains b) => ExampleUni Contains (a, b) where ...

(a universe can have any subset of the mentioned sorts of types, for example it's fine to have instantiated polymorphic types and not have uninstantiated ones and vice versa)

Note that when used as a constraint of a function Contains does not allow you to directly express things like "uni has the Integer, Bool and [] types and type formers", because [] is not fully instantiated. So you can only say "uni has Integer, Bool, [Integer], [Bool], [[Integer]], [[Bool]] etc" and such manual enumeration is annoying, so we'd really like to be able to say that uni has lists of arbitrary built-in types (including lists of lists etc). Contains does not allow that, but Includes does. For example, in the body of the following definition:

foo :: (uni Includes Integer, uni Includes Bool, uni Includes []) => ... foo = ...

you can make use of the fact that uni has lists of arbitrary included types (integers, booleans and lists).

Hence most of the time opt for using the more flexible Includes.

Includes is defined in terms of Contains, so you only need to provide a Contains instance per type from the universe and you'll get Includes for free.

Methods

knownUni :: uni (Esc a)

class Closed (uni :: Type -> Type) where

A universe is Closed, if it's known how to constrain every type from the universe and every type can be encoded to / decoded from a sequence of integer tags. The universe doesn't have to be finite and providing support for infinite universes is the reason why we encode a type as a sequence of integer tags as opposed to a single integer tag. For example, given

  data U a where
      UList :: !(U a) -> U [a]
      UInt  :: U Int

UList (UList UInt) can be encoded to [0,0,1] where 0 and 1 are the integer tags of the UList and UInt constructors, respectively.

Associated Types

type Everywhere (uni :: Type -> Type) (constr :: Type -> Constraint)

A constrant for "constr a holds for any a from uni".

Methods

encodeUni :: uni a -> [Int]

Encode a type as a sequence of Int tags. The opposite of decodeUni.

withDecodedUni :: (forall k (a :: k). Typeable k => uni (Esc a) -> DecodeUniM r) -> DecodeUniM r

Decode a type and feed it to the continuation.

bring :: Everywhere uni constr => proxy constr -> uni (Esc a) -> (constr a => r) -> r

Bring a constr a instance in scope, provided a is a type from the universe and constr holds for any type from the universe.

type family EverywhereAll (uni :: Type -> Type) (constrs :: [Type -> Constraint]) where ...

Equations

EverywhereAll uni ('[] :: [Type -> Constraint]) = ()

EverywhereAll uni (constr ': constrs) = (Everywhere uni constr, EverywhereAll uni constrs)

knownUniOf :: forall {k} uni (a :: k) proxy. Contains uni a => proxy a -> uni (Esc a)

Same as knownUni, but receives a proxy.

class GShow (t :: k -> Type) where

Methods

gshowsPrec :: forall (a :: k). Int -> t a -> ShowS

class GEq (f :: k -> Type) where

Methods

geq :: forall (a :: k) (b :: k). f a -> f b -> Maybe (a :~: b)

class HasUniApply (uni :: Type -> Type) where

A class for "uni has general type application".

Methods

uniApply :: forall k l (f :: k -> l) (a :: k). uni (Esc f) -> uni (Esc a) -> uni (Esc (f a))

Apply a type constructor to an argument.

matchUniApply

Arguments

:: uni tb
The type.

-> r
What to return if the type is not an application.

-> (forall k l (f :: k -> l) (a :: k). tb ~ Esc (f a) => uni (Esc f) -> uni (Esc a) -> r)
The continuation taking a function and an argument.

-> r

Deconstruct a type application into the function and the argument and feed them to the continuation. If the type is not an application, then return the default value.

checkStar :: forall uni a (x :: a). Typeable a => uni (Esc x) -> Maybe (a :~: Type)

Check if the kind of the given type from the universe is Type.

withApplicable :: forall a ab (f :: ab) (x :: a) uni m r. (Typeable ab, Typeable a, MonadPlus m) => uni (Esc f) -> uni (Esc x) -> (forall b. (Typeable b, ab ~ (a -> b)) => m r) -> m r

Check if one type from the universe can be applied to another (i.e. check that the expected kind of the argument matches the actual one) and call the continuation in the refined context. Fail with mzero otherwise.

type (<:) (uni1 :: Type -> Type) (uni2 :: Type -> Type) = Everywhere uni1 (Includes uni2 :: Type -> Constraint)

A constraint for "uni1 is a subuniverse of uni2".

data DefaultUni a where

The universe used by default.

Constructors

DefaultUniInteger :: DefaultUni (Esc Integer)

DefaultUniByteString :: DefaultUni (Esc ByteString)

DefaultUniString :: DefaultUni (Esc Text)

DefaultUniUnit :: DefaultUni (Esc ())

DefaultUniBool :: DefaultUni (Esc Bool)

DefaultUniProtoArray :: DefaultUni (Esc Vector)

DefaultUniProtoList :: DefaultUni (Esc [])

DefaultUniProtoPair :: DefaultUni (Esc (,))

DefaultUniApply :: forall {k1} {k2} (f :: k1 -> k2) (a1 :: k1). !(DefaultUni (Esc f)) -> !(DefaultUni (Esc a1)) -> DefaultUni (Esc (f a1))

DefaultUniData :: DefaultUni (Esc Data)

DefaultUniBLS12_381_G1_Element :: DefaultUni (Esc Element)

DefaultUniBLS12_381_G2_Element :: DefaultUni (Esc Element)

DefaultUniBLS12_381_MlResult :: DefaultUni (Esc MlResult)

DefaultUniValue :: DefaultUni (Esc Value)

pattern DefaultUniList :: forall {a} {k1} {k2} {f :: k1 -> k2} {a1 :: k1}. () => forall. (a ~ Esc (f a1), Esc f ~ Esc []) => DefaultUni (Esc a1) -> DefaultUni a

pattern DefaultUniPair :: forall {a} {k1} {k2} {f1 :: k1 -> k2} {a1 :: k1} {k3} {k4} {f2 :: k3 -> k4} {a2 :: k3}. () => forall. (a ~ Esc (f1 a1), Esc f1 ~ Esc (f2 a2), Esc f2 ~ Esc (,)) => DefaultUni (Esc a2) -> DefaultUni (Esc a1) -> DefaultUni a

pattern DefaultUniArray :: forall {a} {k1} {k2} {f :: k1 -> k2} {a1 :: k1}. () => forall. (a ~ Esc (f a1), Esc f ~ Esc Vector) => DefaultUni (Esc a1) -> DefaultUni a

type family Everywhere (uni :: Type -> Type) (constr :: Type -> Constraint)

A constrant for "constr a holds for any a from uni".

defaultUniSize :: forall k (a :: k). DefaultUni (Esc a) -> Int

noMoreTypeFunctions :: forall a b c d (f :: a -> b -> c -> d) any. DefaultUni (Esc f) -> any

For pleasing the coverage checker.

minBoundInteger :: Integer

type Includes (uni :: Type -> Type) = Permits (Contains uni :: Type -> Constraint) :: k -> Constraint

uni Includes a reads as "a is in the uni". a can be of a higher-kind, see the docs of Contains on why you might want that.

someType :: forall k (a :: k) (uni :: Type -> Type). Contains uni a => SomeTypeIn uni

Wrap a type into SomeTypeIn, provided it's in the universe.

newtype DecodeUniM a

A monad to decode types from a universe in. We use a monad for decoding, because parsing arguments of polymorphic built-in types can peel off an arbitrary amount of type tags from the input list of tags and so we have state, which is convenient to handle with, well, StateT.

Constructors

DecodeUniM

Fields

unDecodeUniM :: StateT [Int] Maybe a

decodeKindedUni :: forall (uni :: Type -> Type). Closed uni => [Int] -> Maybe (SomeTypeIn (Kinded uni))

Decode a type from a sequence of Int tags. The opposite of encodeUni (modulo invalid input).

peelUniTag :: DecodeUniM Int

Peel off a tag from the input list of type tags.

type family Permits (constr :: Type -> Constraint) :: k -> Constraint

constr Permits f elaborates to one of - constr f forall a. constr a => constr (f a) forall a b. (constr a, constr b) => constr (f a b) forall a b c. (constr a, constr b, constr c) => constr (f a b c)

depending on the kind of f. This allows us to say things like

( constr Permits Integer , constr Permits [] , constr Permits (,) )

and thus constraint every type from the universe (including polymorphic ones) to satisfy constr, which is how we provide an implementation of Everywhere for universes with polymorphic types.

Permits is an open type family, so you can provide type instances for fs expecting more type arguments than 3 if you need that.

Note that, say, constr Permits [] elaborates to

forall a. constr a => constr [a]

and for certain type classes that does not make sense (e.g. the Generic instance of [] does not require the type of elements to be Generic), however it's not a problem because we use Permit to constrain the whole universe and so we know that arguments of polymorphic built-in types are builtins themselves are hence do satisfy the constraint and the fact that these constraints on arguments do not get used in the polymorphic case only means that they get ignored.

tryUniApply :: forall m (uni :: Type -> Type). (MonadPlus m, HasUniApply uni) => SomeTypeIn (Kinded uni) -> SomeTypeIn (Kinded uni) -> m (SomeTypeIn (Kinded uni))

Apply a type constructor to an argument, provided kinds match.

gshow :: forall {k} t (a :: k). GShow t => t a -> String

defaultEq :: forall {k} f (a :: k) (b :: k). GEq f => f a -> f b -> Bool

data DSum (tag :: k -> Type) (f :: k -> Type)

Constructors

!(tag a) :=> (f a)

maxBoundInteger :: Integer

maxBoundByteString :: Int

newtype CInteger

Cardano on-chain integer

Constructors

CInteger Integer

newtype CByteString

Cardano on-chain bytestring

Constructors

CByteString ByteString