PlutusCore.Pretty

Basic types and functions

data Doc ann

class Pretty a where

Minimal complete definition

pretty

Methods

pretty :: a -> Doc ann

prettyList :: [a] -> Doc ann

class PrettyBy config a where

Minimal complete definition

Nothing

Methods

prettyBy :: config -> a -> Doc ann

prettyListBy :: config -> [a] -> Doc ann

newtype IgnorePrettyConfig a

Constructors

IgnorePrettyConfig

Fields

unIgnorePrettyConfig :: a

data AttachPrettyConfig config a

Constructors

AttachPrettyConfig !config !a

class Render str where

Methods

render :: Doc ann -> str

type PrettyParens = PrettyBy RenderContext

For pretty-printing a value with a minimum amount of parens.

display :: forall str a. (Pretty a, Render str) => a -> str

displayBy :: forall str a config. (PrettyBy config a, Render str) => config -> a -> str

juxtRenderContext :: RenderContext

An initial RenderContext. An expression printed in this context gets enclosed in parens unless its outermost operator (if any) binds even stronger than function application.

Defaults

prettyPlc :: PrettyPlc a => a -> Doc ann

Pretty-print a value in the default mode using the classic view.

displayPlc :: (PrettyPlc a, Render str) => a -> str

Render a value to String in the default mode using the classic view.

prettyPlcSimple :: PrettyPlc a => a -> Doc ann

Pretty-print a value in the debug mode using the classic view.

displayPlcSimple :: (PrettyPlc a, Render str) => a -> str

Render a value to String in the debug mode using the classic view.

Global configuration

data CondensedErrors

Whether to pretty-print PLC errors in full or with some information omitted.

Constructors

CondensedErrorsYes

CondensedErrorsNo

type DefaultPrettyPlcStrategy a = (PrettyClassic a, PrettyReadable a)

A constraint that allows to derive PrettyBy PrettyConfigPlc instances, see below.

newtype PrettyConfigPlcOptions

Options for pretty-printing PLC entities.

Constructors

PrettyConfigPlcOptions

Fields

_pcpoCondensedErrors :: CondensedErrors

data PrettyConfigPlcStrategy

Strategy for pretty-printing PLC entities.

Constructors

PrettyConfigPlcClassic (PrettyConfigClassic PrettyConfigName)

PrettyConfigPlcReadable (PrettyConfigReadable PrettyConfigName)

data PrettyConfigPlc

Global configuration used for pretty-printing PLC entities.

Constructors

PrettyConfigPlc

Fields

_pcpOptions :: PrettyConfigPlcOptions

_pcpStrategy :: PrettyConfigPlcStrategy

type PrettyPlc = PrettyBy PrettyConfigPlc

The "pretty-printable PLC entity" constraint.

prettyConfigPlcOptions :: PrettyConfigPlcOptions

The PrettyConfigPlcOptions used by default: print errors in full.

prettyConfigPlcClassic :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used by default: use the classic view and print neither Uniques, nor name attachments.

prettyConfigPlcClassicSimple :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used for debugging: use the classic view and print Uniques, but not name attachments.

prettyConfigPlcReadable :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used by default and for readability: use the refined view and print Uniques but not name attachments.

prettyConfigPlcReadableSimple :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used for debugging and readability: use the refined view and print neither Uniques nor name attachments.

Custom functions for PLC types.

prettyPlcClassic :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the default mode using the classic view.

prettyPlcClassicSimple :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value without unique indices using the classic view.

prettyPlcReadable :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the default mode using the readable view.

prettyPlcReadableSimple :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value without unique indices using the readable view.

prettyPlcCondensedErrorBy :: PrettyPlc a => (PrettyConfigPlcOptions -> PrettyConfigPlc) -> a -> Doc ann

Pretty-print a PLC value using the condensed way (see CondensedErrors) of pretty-printing PLC errors (in case there are any).

displayPlcCondensedErrorClassic :: (PrettyPlc a, Render str) => a -> str

Render an error to String in the condensed manner using the classic view.

Names

newtype PrettyConfigName

A config that determines how to pretty-print a PLC name.

Constructors

PrettyConfigName

Fields

_pcnShowsUnique :: Bool

Whether to show the Unique of a name or not.

class HasPrettyConfigName config where

A class of configs from which a PrettyConfigName can be extracted.

Methods

toPrettyConfigName :: config -> PrettyConfigName

prettyConfigName :: PrettyConfigName

The PrettyConfigName used by default: print Unique indexes after nams.

prettyConfigNameSimple :: PrettyConfigName

The PrettyConfigName to be used when Unique indices don't matter. Easier to read.

Classic view

data PrettyConfigClassic configName

Configuration for the classic pretty-printing.

Constructors

PrettyConfigClassic

Fields

_pccConfigName :: configName

How to pretty-print names.

_pccDisplayAnn :: Bool

Whether to display annotations.

type PrettyClassicBy configName = PrettyBy (PrettyConfigClassic configName)

The "classically pretty-printable" constraint.

type PrettyClassic = PrettyClassicBy PrettyConfigName

consAnnIf :: Pretty ann => PrettyConfigClassic configName -> ann -> [Doc dann] -> [Doc dann]

Add a pretty-printed annotation to a list of Docs if the given config enables pretty-printing of annotations.

prettyClassic :: PrettyClassic a => a -> Doc ann

Pretty-print a value in the default mode using the classic view.

prettyClassicSimple :: PrettyClassic a => a -> Doc ann

Pretty-print a value in the simple mode using the classic view.

Readable view

data ShowKinds

Constructors

ShowKindsYes

ShowKindsNonType

ShowKindsNo

data PrettyConfigReadable configName

Configuration for the readable pretty-printing.

Constructors

PrettyConfigReadable

Fields

_pcrConfigName :: configName

_pcrRenderContext :: RenderContext

_pcrShowKinds :: ShowKinds

prettyReadable :: PrettyReadable a => a -> Doc ann

Pretty-print something with the PrettyConfigReadable config.

prettyReadableSimple :: PrettyReadable a => a -> Doc ann

Pretty-print something with the PrettyConfigReadableSimple config.

pcrConfigName :: forall configName1 configName2 f. Functor f => (configName1 -> f configName2) -> PrettyConfigReadable configName1 -> f (PrettyConfigReadable configName2)

pcrRenderContext :: forall configName f. Functor f => (RenderContext -> f RenderContext) -> PrettyConfigReadable configName -> f (PrettyConfigReadable configName)

pcrShowKinds :: forall configName f. Functor f => (ShowKinds -> f ShowKinds) -> PrettyConfigReadable configName -> f (PrettyConfigReadable configName)

type PrettyReadableBy configName = PrettyBy (PrettyConfigReadable configName)

The "readably pretty-printable" constraint.

type PrettyReadable = PrettyReadableBy PrettyConfigName

newtype AsReadable a

For rendering things in a readable manner regardless of the pretty-printing function chosen. I.e. all of show, pretty, prettyClassic will use PrettyReadable instead of doing what they normally do. prettyBy config (AsReadable x) requires config to have a PrettyConfigName and respects it.

This wrapper can be particularly useful if you want to apply a function having a Show or Pretty or PrettyClassic or PrettyPlc or whatever constraint, but want to get the argument rendered in a readable manner instead.

Constructors

AsReadable

Fields

unAsReadable :: a

data Parened a

A value of type a to render in parens using the readable pretty-printer.

Constructors

Parened

Fields

parenOpening :: String

parenClosing :: String

parenedValue :: a

inBraces :: a -> Parened a

Enclose the given value, so that it's rendered inside of braces with no additional parens regardless of the RenderContext.

topPrettyConfigReadable :: configName -> ShowKinds -> PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to topFixity.

botPrettyConfigReadable :: configName -> ShowKinds -> PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to botFixity.

binderFixity :: Fixity

The fixity of a binder.

arrowFixity :: Fixity

The fixity of (->).

iterTyForallPrettyM :: (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) => [arg] -> body -> m (Doc ann)

Lay out an iterated TyForall via iterBinderPrettyM.

iterLamAbsPrettyM :: (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) => [arg] -> body -> m (Doc ann)

Lay out an iterated LamAbs via iterBinderPrettyM.

iterTyAbsPrettyM :: (MonadPrettyReadable configName env m, PrettyReadableBy configName arg, PrettyReadableBy configName body) => [arg] -> body -> m (Doc ann)

Lay out an iterated TyAbs via iterBinderPrettyM.

iterArrowPrettyM :: (MonadPrettyReadable configName env m, PrettyReadableBy configName a) => [a] -> a -> m (Doc ann)

Lay out an iterated ->.

iterAppDocM :: MonadPrettyContext config env m => (AnyToDoc config ann -> AnyToDoc config ann -> NonEmpty (Doc ann)) -> m (Doc ann)

Lay out an iteration application, providing to the caller a function to render the head of the application and a function to render each of the arguments.

iterInterAppPrettyM :: (MonadPrettyReadable configName env m, PrettyReadableBy configName fun, PrettyReadableBy configName ty, PrettyReadableBy configName term) => fun -> [Either ty term] -> m (Doc ann)

Lay out interleaved function applications either as

foo {a} x {b} y z

or as

foo
  {a}
  x
  {b}
  y
  z

Lefts are laid out in braces, Rights are laid out without braces.

iterAppPrettyM :: (MonadPrettyContext config env m, PrettyBy config fun, PrettyBy config term) => fun -> [term] -> m (Doc ann)

Lay out iterated function applications either as

foo x y z

or as

foo
  x
  y
  z 

Utils

prettyBytes :: ByteString -> Doc ann

newtype ConstConfig

The type of configs used for pretty-printing constants. Has a RenderContext inside, so that we don't add redundant parens to the output.

Constructors

ConstConfig

Fields

unConstConfig :: RenderContext

type PrettyConst = PrettyBy ConstConfig

type PrettyUni (uni :: Type -> Type) = (PrettyParens (SomeTypeIn uni), Closed uni, Everywhere uni PrettyConst)

The set of constraints we need to be able to print built-in types and their values.

type ThrowableBuiltins (uni :: Type -> Type) fun = (PrettyUni uni, Pretty fun, Typeable uni, Typeable fun)

The set of constraints we need to be able to throw exceptions with things with built-in types and functions in them.

prettyConst :: PrettyConst a => RenderContext -> a -> Doc ann

data Associativity

Constructors

LeftAssociative

RightAssociative

NonAssociative

type Fixity = FixityOver Precedence

data FixityOver prec

Constructors

Fixity

Fields

_fixityAssociativity :: !Associativity

_fixityPrecedence :: !prec

type Precedence = Double

type RenderContext = RenderContextOver Precedence

data RenderContextOver prec

Constructors

RenderContext

Fields

_renderContextDirection :: !Direction

_renderContextFixity :: !(FixityOver prec)

inContextM :: (a -> InContextM config (Doc ann)) -> config -> a -> Doc ann

juxtPrettyM :: (MonadPrettyContext config env m, PrettyBy config a, PrettyBy config b) => a -> b -> m (Doc ann)

newtype Sole a

Constructors

Sole

Fields

unSole :: a

botFixity :: Fixity

botRenderContext :: RenderContext

juxtFixity :: Fixity

topFixity :: Fixity

topRenderContext :: RenderContext

unitFixity :: Fixity

data Direction

Constructors

ToTheLeft

ToTheRight

encloseIn :: Ord prec => (a -> a) -> RenderContextOver prec -> FixityOver prec -> a -> a

class HasPrettyConfig env config | env -> config where

Methods

prettyConfig :: Lens' env config

type MonadPretty config env (m :: Type -> Type) = (MonadReader env m, HasPrettyConfig env config)

displayM :: forall str a m env config. (MonadPretty config env m, PrettyBy config a, Render str) => a -> m str

prettyM :: (MonadPretty config env m, PrettyBy config a) => a -> m (Doc ann)

class HasRenderContext config where

Methods

renderContext :: Lens' config RenderContext

newtype InContextM config a

Constructors

InContextM

Fields

unInContextM :: Reader (Sole config) a

compoundDocM :: MonadPrettyContext config env m => Fixity -> ((forall a. PrettyBy config a => Direction -> Fixity -> a -> Doc ann) -> Doc ann) -> m (Doc ann)

encloseM :: MonadPrettyContext config env m => Fixity -> Doc ann -> m (Doc ann)

infixDocM :: MonadPrettyContext config env m => Fixity -> (AnyToDoc config ann -> AnyToDoc config ann -> Doc ann) -> m (Doc ann)

runInContextM :: config -> InContextM config a -> a

sequenceDocM :: MonadPrettyContext config env m => Direction -> Fixity -> (AnyToDoc config ann -> Doc ann) -> m (Doc ann)

unitDocM :: MonadPrettyContext config env m => Doc ann -> m (Doc ann)

withPrettyAt :: MonadPrettyContext config env m => Direction -> Fixity -> (AnyToDoc config ann -> m r) -> m r

withPrettyIn :: MonadPrettyContext config env m => ((forall a. PrettyBy config a => Direction -> Fixity -> a -> Doc ann) -> m r) -> m r

type AnyToDoc config ann = forall a. PrettyBy config a => a -> Doc ann

type MonadPrettyContext config env (m :: Type -> Type) = (MonadPretty config env m, HasRenderContext config)