UntypedPlutusCore

Documentation

module PlutusCore.Name.Unique

module UntypedPlutusCore.AstSize

module UntypedPlutusCore.Check.Scope

module UntypedPlutusCore.Core

module UntypedPlutusCore.DeBruijn

data SimplifierStage

Constructors

FloatDelay

ForceDelay

ForceCaseDelay

CaseOfCase

CaseReduce

Inline

CSE

ApplyToCase

Unknown

termMapNames :: forall name name' (uni :: Type -> Type) fun ann. (name -> name') -> Term name uni fun ann -> Term name' uni fun ann

Completely replace the names with a new name type.

programMapNames :: forall name name' (uni :: Type -> Type) fun ann. (name -> name') -> Program name uni fun ann -> Program name' uni fun ann

newtype InlineHints name a

Constructors

InlineHints

Fields

shouldInline :: a -> name -> Inline

data SimplifyOpts name a

Constructors

SimplifyOpts

Fields

_soMaxSimplifierIterations :: Int

_soMaxCseIterations :: Int

_soCseWhichSubterms :: CseWhichSubterms

_soConservativeOpts :: Bool

_soInlineHints :: InlineHints name a

_soInlineConstants :: Bool

_soInlineCallsiteGrowth :: AstSize

_soPreserveLogging :: Bool

_soApplyToCase :: Bool

defaultSimplifyOpts :: SimplifyOpts name a

substVarA :: forall f name ann (uni :: Type -> Type) fun. Applicative f => (name -> ann -> f (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> f (Term name uni fun ann)

Applicatively replace a variable using the given function.

substVar :: forall name ann (uni :: Type -> Type) fun. (name -> ann -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Replace a variable using the given function.

termSubstNamesM :: forall m name ann (uni :: Type -> Type) fun. Monad m => (name -> ann -> m (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> m (Term name uni fun ann)

Naively monadically substitute names using the given function (i.e. do not substitute binders).

termSubstNames :: forall name ann (uni :: Type -> Type) fun. (name -> ann -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Naively substitute names using the given function (i.e. do not substitute binders).

vTerm :: forall name (uni :: Type -> Type) fun ann f. (Contravariant f, Applicative f) => (name -> f name) -> Term name uni fun ann -> f (Term name uni fun ann)

Get all the term variables in a term.

substConstantA :: forall f ann (uni :: Type -> Type) name fun. Applicative f => (ann -> Some (ValueOf uni) -> f (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> f (Term name uni fun ann)

Applicatively replace a constant using the given function.

substConstant :: forall ann (uni :: Type -> Type) name fun. (ann -> Some (ValueOf uni) -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Replace a constant using the given function.

termSubstConstantsM :: forall m ann (uni :: Type -> Type) name fun. Monad m => (ann -> Some (ValueOf uni) -> m (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> m (Term name uni fun ann)

Monadically substitute constants using the given function.

termSubstConstants :: forall ann (uni :: Type -> Type) name fun. (ann -> Some (ValueOf uni) -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Substitute constants using the given function.

newtype SimplifierT name (uni :: Type -> Type) fun ann (m :: Type -> Type) a

Constructors

SimplifierT

Fields

getSimplifierT :: StateT (SimplifierTrace name uni fun ann) m a

recordSimplification :: forall (m :: Type -> Type) name (uni :: Type -> Type) fun a. Monad m => Term name uni fun a -> SimplifierStage -> Term name uni fun a -> SimplifierT name uni fun a m ()

type Simplifier name (uni :: Type -> Type) fun ann = SimplifierT name uni fun ann Identity

data CseWhichSubterms

Which subterms should be considered as candidates?

Constructors

AllSubterms

ExcludeWorkFree

recordSimplificationWithHints :: forall (m :: Type -> Type) name (uni :: Type -> Type) fun a. Monad m => Hints -> Term name uni fun a -> SimplifierStage -> Term name uni fun a -> SimplifierT name uni fun a m ()

newtype SimplifierTrace name (uni :: Type -> Type) fun a

Constructors

SimplifierTrace

Fields

simplifierTrace :: [Simplification name uni fun a]

data Simplification name (uni :: Type -> Type) fun a

Constructors

Simplification

Fields

beforeAST :: Term name uni fun a

stage :: SimplifierStage

hints :: Hints

afterAST :: Term name uni fun a

runSimplifierT :: forall name (uni :: Type -> Type) fun ann m a. SimplifierT name uni fun ann m a -> m (a, SimplifierTrace name uni fun ann)

evalSimplifierT :: forall m name (uni :: Type -> Type) fun ann a. Monad m => SimplifierT name uni fun ann m a -> m a

execSimplifierT :: forall m name (uni :: Type -> Type) fun ann a. Monad m => SimplifierT name uni fun ann m a -> m (SimplifierTrace name uni fun ann)

runSimplifier :: forall name (uni :: Type -> Type) fun ann a. Simplifier name uni fun ann a -> (a, SimplifierTrace name uni fun ann)

evalSimplifier :: forall name (uni :: Type -> Type) fun ann a. Simplifier name uni fun ann a -> a

execSimplifier :: forall name (uni :: Type -> Type) fun ann a. Simplifier name uni fun ann a -> SimplifierTrace name uni fun ann

initSimplifierTrace :: forall name (uni :: Type -> Type) fun a. SimplifierTrace name uni fun a

simplifyTerm :: forall name (uni :: Type -> Type) fun m a. Compiling m uni fun name a => SimplifyOpts name a -> BuiltinSemanticsVariant fun -> Term name uni fun a -> m (Term name uni fun a)

simplifyProgram :: forall name (uni :: Type -> Type) fun m a. Compiling m uni fun name a => SimplifyOpts name a -> BuiltinSemanticsVariant fun -> Program name uni fun a -> m (Program name uni fun a)

simplifyProgramWithTrace :: forall name (uni :: Type -> Type) fun m a. Compiling m uni fun name a => SimplifyOpts name a -> BuiltinSemanticsVariant fun -> Program name uni fun a -> m (Program name uni fun a, SimplifierTrace name uni fun a)

termSimplifier :: forall name (uni :: Type -> Type) fun (m :: Type -> Type) a. Compiling m uni fun name a => SimplifyOpts name a -> BuiltinSemanticsVariant fun -> Term name uni fun a -> SimplifierT name uni fun a m (Term name uni fun a)

soApplyToCase :: forall name a f. Functor f => (Bool -> f Bool) -> SimplifyOpts name a -> f (SimplifyOpts name a)

soMaxSimplifierIterations :: forall name a f. Functor f => (Int -> f Int) -> SimplifyOpts name a -> f (SimplifyOpts name a)

soMaxCseIterations :: forall name a f. Functor f => (Int -> f Int) -> SimplifyOpts name a -> f (SimplifyOpts name a)

soCseWhichSubterms :: forall name a f. Functor f => (CseWhichSubterms -> f CseWhichSubterms) -> SimplifyOpts name a -> f (SimplifyOpts name a)

soInlineHints :: forall name1 a1 name2 a2 f. Functor f => (InlineHints name1 a1 -> f (InlineHints name2 a2)) -> SimplifyOpts name1 a1 -> f (SimplifyOpts name2 a2)

soConservativeOpts :: forall name a f. Functor f => (Bool -> f Bool) -> SimplifyOpts name a -> f (SimplifyOpts name a)

soInlineConstants :: forall name a f. Functor f => (Bool -> f Bool) -> SimplifyOpts name a -> f (SimplifyOpts name a)

soInlineCallsiteGrowth :: forall name a f. Functor f => (AstSize -> f AstSize) -> SimplifyOpts name a -> f (SimplifyOpts name a)

soPreserveLogging :: forall name a f. Functor f => (Bool -> f Bool) -> SimplifyOpts name a -> f (SimplifyOpts name a)

data Term name (uni :: Type -> Type) fun ann

The type of Untyped Plutus Core terms. Mirrors the type of Typed Plutus Core terms except

all types are removed

IWrap and Unwrap are removed

type abstractions are replaced with Delay

type instantiations are replaced with Force

The latter two are due to the fact that we don't have value restriction in Typed Plutus Core and hence a computation can be stuck expecting only a single type argument for the computation to become unstuck. Therefore we can't just silently remove type abstractions and instantiations and need to replace them with something else that also blocks evaluation (in order for the semantics of an erased program to match with the semantics of the original typed one). Delay and Force serve exactly this purpose.

Constructors

Var !ann !name

LamAbs !ann !name !(Term name uni fun ann)

Apply !ann !(Term name uni fun ann) !(Term name uni fun ann)

Force !ann !(Term name uni fun ann)

Delay !ann !(Term name uni fun ann)

Constant !ann !(Some (ValueOf uni))

Builtin !ann !fun

Error !ann

Constr !ann !Word64 ![Term name uni fun ann]

Case !ann !(Term name uni fun ann) !(Vector (Term name uni fun ann))

data Program name (uni :: Type -> Type) fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn :: ann

_progVer :: Version

_progTerm :: Term name uni fun ann

applyProgram :: forall m a name (uni :: Type -> Type) fun. (MonadError ApplyProgramError m, Semigroup a) => Program name uni fun a -> Program name uni fun a -> m (Program name uni fun a)

Applies one program to another. Fails if the versions do not match and tries to merge annotations.

parseScoped :: forall (uni :: Type -> Type) fun m. (MonadError (Error uni fun SrcSpan) m, MonadQuote m) => Text -> m (Program Name DefaultUni DefaultFun SrcSpan)

Parse and rewrite so that names are globally unique, not just unique within their scope.

data DefaultUni a

The universe used by default.

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins/language versions and protocol versions]